Finish server route extraction and harden persona review validation
