+++ apps/server/src/server.ts
@@ -134,16 +134,21 @@ import {
   releaseNotesMarkdown,
   staticFiles as embeddedStaticFiles,
 } from "./generated/runtime-assets.js";
-import {
-  computeActivityStats,
-  computeDailyStatus,
-  computeWorkingTimeByProject,
-  type ActivityEventRow,
-} from "./activity-metrics.js";
+import { type ActivityEventRow } from "./activity-metrics.js";
+import { registerActivityRoutes } from "./routes/activity.js";
+import { registerAgentRoutes } from "./routes/agents.js";
 import { registerAuthRoutes } from "./routes/auth.js";
 import { registerFeedbackRoutes } from "./routes/feedback.js";
+import { registerJobRoutes } from "./routes/jobs.js";
 import { registerMediaRoutes } from "./routes/media.js";
+import { registerMcpRoutes } from "./routes/mcp.js";
 import { registerPersonaReviewRoutes } from "./routes/persona-reviews.js";
+import {
+  registerReleaseRoutes,
+  type ReleaseJob as RouteReleaseJob,
+} from "./routes/release.js";
+import { registerStaticRoutes } from "./routes/static.js";
+import { registerSystemRoutes } from "./routes/system.js";
 
 const config = loadConfig();
 const app = Fastify({
@@ -1589,43 +1594,9 @@ function invalidatePasswordSetCache(): void {
 
 const SESSION_COOKIE = "dispatch_session";
 const SESSION_MAX_AGE_S = 30 * 24 * 60 * 60; // 30 days
-function resolveTilde(raw: string): string {
-  if (raw.startsWith("~/")) return path.join(os.homedir(), raw.slice(2));
-  if (raw === "~") return os.homedir();
-  return raw;
-}
-const directoryField = z
-  .string()
-  .min(1, "Job directory is required.")
-  .transform(resolveTilde);
-const RunJobBodySchema = z.object({
-  name: z.string().min(1, "Job name is required."),
-  directory: directoryField,
-  wait: z.boolean().optional(),
-});
-const JobEnableDisableBodySchema = z.object({
-  name: z.string().min(1, "Job name is required."),
-  directory: directoryField,
-});
-const AddJobBodySchema = JobEnableDisableBodySchema.extend({
-  displayName: z.string().optional(),
-  prompt: z.string().nullable().optional(),
-  schedule: z.string().nullable().optional(),
-  timeoutMs: z.number().int().positive().optional(),
-  needsInputTimeoutMs: z.number().int().positive().optional(),
-  agentType: z.enum(CLI_AGENT_TYPES).optional(),
-  useWorktree: z.boolean().optional(),
-  baseBranch: z.string().nullable().optional(),
-  branchName: z.string().nullable().optional(),
-  fullAccess: z.boolean().optional(),
-  autoArchive: z.boolean().optional(),
-  enabled: z.boolean().optional(),
-});
-const JobHistoryParamsSchema = z.object({
-  name: z.string().min(1, "Job name is required."),
-  directory: directoryField,
-  limit: z.coerce.number().int().min(1).max(100).optional(),
-});
+const WORKTREE_LOCATION_KEY = "worktree_location";
+type WorktreeLocation = "sibling" | "nested";
+const VALID_WORKTREE_LOCATIONS: WorktreeLocation[] = ["sibling", "nested"];
 
 async function registerRoutes() {
   const cookieSecret = await getOrCreateCookieSecret(pool);
@@ -1650,51 +1621,10 @@ async function registerRoutes() {
     rewriteForColor(storedIconColor as IconColor);
   }
 
-  // Serve templated index.html and manifest before static files so they take priority
-  const noCacheHeaders = {
-    "Cache-Control": "no-cache, no-store, must-revalidate",
-  };
-
-  app.get("/", async (_, reply) => {
-    return reply
-      .type("text/html")
-      .headers(noCacheHeaders)
-      .send(cachedIndexHtml);
-  });
-
-  app.get("/index.html", async (_, reply) => {
-    return reply
-      .type("text/html")
-      .headers(noCacheHeaders)
-      .send(cachedIndexHtml);
-  });
-
-  app.get("/manifest.webmanifest", async (_, reply) => {
-    return reply
-      .type("application/manifest+json")
-      .headers(noCacheHeaders)
-      .send(cachedManifest);
-  });
-
-  // SPA fallback — serve templated index.html for client-side routes
-  app.setNotFoundHandler(async (request, reply) => {
-    const url = request.url.split("?")[0];
-    const staticAsset = staticAssets.get(url);
-    if (staticAsset) {
-      const isServiceWorker = path.basename(url) === "sw.js";
-      const response = reply.type(staticAsset.contentType);
-      if (isServiceWorker) {
-        response.headers(noCacheHeaders);
-      }
-      return response.send(staticAsset.body);
-    }
-    if (!url.startsWith("/api/") && !path.extname(url)) {
-      return reply
-        .type("text/html")
-        .headers(noCacheHeaders)
-        .send(cachedIndexHtml);
-    }
-    return reply.code(404).send({ error: "Not found" });
+  await registerStaticRoutes(app, {
+    cachedIndexHtml,
+    cachedManifest,
+    staticAssets,
   });
 
   // ---------------------------------------------------------------------------
@@ -1759,2421 +1689,116 @@ async function registerRoutes() {
     invalidatePasswordSetCache,
   });
 
-  // ---------------------------------------------------------------------------
-  // Jobs routes
-  // ---------------------------------------------------------------------------
-
-  app.post("/api/v1/jobs/run", async (request, reply) => {
-    const parsed = RunJobBodySchema.safeParse(request.body);
-    if (!parsed.success) {
-      return reply.code(400).send({ error: parsed.error.issues[0].message });
-    }
-
-    try {
-      return await jobService.runJob(parsed.data);
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      return reply.code(500).send({ error: message });
-    }
-  });
-
-  app.get("/api/v1/jobs", async () => {
-    return await jobService.listJobs();
-  });
-
-  app.post("/api/v1/jobs", async (request, reply) => {
-    const parsed = AddJobBodySchema.safeParse(request.body);
-    if (!parsed.success) {
-      return reply.code(400).send({ error: parsed.error.issues[0].message });
-    }
-    try {
-      const result = await jobService.addJob(parsed.data);
-      uiEventBroker.publish({ type: "job.changed" });
-      return result;
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      return reply.code(500).send({ error: message });
-    }
-  });
-
-  app.patch("/api/v1/jobs", async (request, reply) => {
-    const parsed = AddJobBodySchema.safeParse(request.body);
-    if (!parsed.success) {
-      return reply.code(400).send({ error: parsed.error.issues[0].message });
-    }
-    try {
-      const result = await jobService.updateJob(parsed.data);
-      uiEventBroker.publish({ type: "job.changed" });
-      return result;
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      return reply.code(500).send({ error: message });
-    }
-  });
-
-  app.delete("/api/v1/jobs", async (request, reply) => {
-    const parsed = JobEnableDisableBodySchema.safeParse(request.body);
-    if (!parsed.success) {
-      return reply.code(400).send({ error: parsed.error.issues[0].message });
-    }
-    try {
-      const result = await jobService.removeJob(parsed.data);
-      uiEventBroker.publish({ type: "job.changed" });
-      return result;
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      return reply.code(500).send({ error: message });
-    }
-  });
-
-  app.post("/api/v1/jobs/enable", async (request, reply) => {
-    const parsed = JobEnableDisableBodySchema.safeParse(request.body);
-    if (!parsed.success) {
-      return reply.code(400).send({ error: parsed.error.issues[0].message });
-    }
-    try {
-      const result = await jobService.enableJob(parsed.data);
-      uiEventBroker.publish({ type: "job.changed" });
-      return result;
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      return reply.code(500).send({ error: message });
-    }
-  });
-
-  app.post("/api/v1/jobs/disable", async (request, reply) => {
-    const parsed = JobEnableDisableBodySchema.safeParse(request.body);
-    if (!parsed.success) {
-      return reply.code(400).send({ error: parsed.error.issues[0].message });
-    }
-    try {
-      const result = await jobService.disableJob(parsed.data);
-      uiEventBroker.publish({ type: "job.changed" });
-      return result;
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      return reply.code(500).send({ error: message });
-    }
-  });
-
-  app.get("/api/v1/jobs/stats", async (_request, reply) => {
-    try {
-      return await jobService.getStats();
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      return reply.code(500).send({ error: message });
-    }
-  });
-
-  app.get("/api/v1/jobs/history", async (request, reply) => {
-    const parsed = JobHistoryParamsSchema.safeParse(request.query);
-    if (!parsed.success) {
-      return reply.code(400).send({ error: parsed.error.issues[0].message });
-    }
-    try {
-      return await jobService.listRunsForJob(parsed.data);
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      return reply.code(404).send({ error: message });
-    }
-  });
-
-  // ---------------------------------------------------------------------------
-  // MCP routes
-  // ---------------------------------------------------------------------------
-
-  app.post("/api/mcp", async (request, reply) => {
-    reply.hijack();
-    await handleMcpRequest(request.raw, reply.raw, request.body);
-  });
-
-  app.post("/api/mcp/jobs/:runId/:agentId", async (request, reply) => {
-    const params = request.params as { runId?: string; agentId?: string };
-    const runId = params.runId ?? "";
-    const agentId = params.agentId ?? "";
-    const bearerToken = getBearerToken(request);
-    if (
-      bearerToken &&
-      !validateJobMcpToken(config.authToken, bearerToken, runId, agentId)
-    ) {
-      return reply.code(403).send({
-        error: "Invalid MCP token for the requested job agent route.",
-      });
-    }
-    const agent = await agentManager.getAgent(agentId);
-    if (!agent) {
-      return reply.code(404).send({ error: "Agent not found." });
-    }
-    const run = await jobService.getActiveRunForAgent(agentId);
-    if (!run || run.id !== runId || run.agentId !== agentId) {
-      return reply
-        .code(404)
-        .send({ error: "Active job run not found for agent." });
-    }
-
-    let repoRoot: string | null = null;
-    let worktreeRoot: string | null = null;
-    try {
-      repoRoot = await resolveRepoRoot(agent.cwd);
-      worktreeRoot = await resolveWorktreeRoot(agent.cwd);
-    } catch {
-      // Agent may not be in a git repository — MCP still works, just without repo context.
-    }
-
-    reply.hijack();
-    await handleMcpRequest(request.raw, reply.raw, request.body, {
-      agent: {
-        id: agent.id,
-        cwd: agent.cwd,
-        persona: agent.persona,
-        parentAgentId: agent.parentAgentId,
-        baseBranch: agent.baseBranch,
-        review: null,
-      },
-      repoRoot,
-      worktreeRoot,
-      sendNotify: mcpSendNotify,
-      upsertEvent: mcpUpsertEvent,
-      renameSession: mcpRenameSession,
-      shareMedia: mcpShareMedia,
-      upsertPin: mcpUpsertPin,
-      deletePin: mcpDeletePin,
-      toolScope: "job",
-      jobTools: {
-        complete: mcpJobComplete,
-        failed: mcpJobFailed,
-        needsInput: mcpJobNeedsInput,
-        log: mcpJobLog,
-        listAgents: async () => {
-          const agents = await agentManager.listAgents();
-          return agents.map((a) => ({
-            id: a.id,
-            name: a.name,
-            status: a.status,
-            cwd: a.cwd,
-          }));
-        },
-        listRecentPersonaReviews: (sinceDays: number) =>
-          agentManager.listRecentPersonaReviews(sinceDays),
-        listRecentFeedback: (sinceDays: number) =>
-          agentManager.listRecentFeedback(sinceDays),
-        getActivitySummary: (params) => agentManager.getActivitySummary(params),
-        getAgentHistory: (params) => agentManager.getAgentHistory(params),
-        getFeedbackSummary: (params) => agentManager.getFeedbackSummary(params),
-      },
-    });
-  });
-
-  app.post("/api/mcp/:agentId", async (request, reply) => {
-    const params = request.params as { agentId?: string };
-    const agentId = params.agentId ?? "";
-    const bearerToken = getBearerToken(request);
-    if (
-      bearerToken &&
-      !validateAgentMcpToken(config.authToken, bearerToken, agentId)
-    ) {
-      return reply
-        .code(403)
-        .send({ error: "Invalid MCP token for the requested agent route." });
-    }
-    const agent = await agentManager.getAgent(agentId);
-    if (!agent) {
-      return reply.code(404).send({ error: "Agent not found." });
-    }
-    const review = agent.persona
-      ? await agentManager.getPersonaReview(agentId)
-      : null;
-    const activeJobRun = await jobService.getActiveRunForAgent(agentId);
-    if (activeJobRun) {
-      return reply
-        .code(403)
-        .send({ error: "Job agents must use the job-scoped MCP route." });
-    }
-
-    let repoRoot: string | null = null;
-    let worktreeRoot: string | null = null;
-    try {
-      repoRoot = await resolveRepoRoot(agent.cwd);
-      worktreeRoot = await resolveWorktreeRoot(agent.cwd);
-    } catch {
-      // Agent may not be in a git repository — MCP still works, just without repo context.
-    }
-
-    reply.hijack();
-    await handleMcpRequest(request.raw, reply.raw, request.body, {
-      agent: {
-        id: agent.id,
-        cwd: agent.cwd,
-        persona: agent.persona,
-        parentAgentId: agent.parentAgentId,
-        baseBranch: agent.baseBranch,
-        review: review
-          ? { allowRecheck: review.allowRecheck, status: review.status }
-          : null,
-      },
-      repoRoot,
-      worktreeRoot,
-      sendNotify: mcpSendNotify,
-      upsertEvent: mcpUpsertEvent,
-      renameSession: mcpRenameSession,
-      shareMedia: mcpShareMedia,
-      listMedia: mcpListMedia,
-      submitFeedback: mcpSubmitFeedback,
-      listPersonas: mcpListPersonas,
-      launchPersona: mcpLaunchPersona,
-      getFeedback: mcpGetFeedback,
-      resolveFeedback: mcpResolveFeedback,
-      submitResolution: mcpSubmitResolution,
-      cancelRecheck: mcpCancelRecheck,
-      upsertPin: mcpUpsertPin,
-      deletePin: mcpDeletePin,
-      getParentContext: mcpGetParentContext,
-      getRecheckContext: mcpGetRecheckContext,
-      updateReviewStatus: mcpUpdateReviewStatus,
-      completeReview: mcpCompleteReview,
-      getActivitySummary: (params) =>
-        agentManager.getActivitySummary(params) as Promise<
-          Record<string, unknown>
-        >,
-      getAgentHistory: (params) =>
-        agentManager.getAgentHistory(params) as Promise<
-          Record<string, unknown>
-        >,
-      getFeedbackSummary: (params) =>
-        agentManager.getFeedbackSummary(params) as Promise<
-          Record<string, unknown>
-        >,
-    });
-  });
-
-  app.get("/api/mcp", async (_, reply) => {
-    return reply.code(405).send(mcpMethodNotAllowed());
-  });
-
-  app.delete("/api/mcp", async (_, reply) => {
-    return reply.code(405).send(mcpMethodNotAllowed());
-  });
-
-  app.get("/api/mcp/:agentId", async (_, reply) => {
-    return reply.code(405).send(mcpMethodNotAllowed());
-  });
-
-  app.delete("/api/mcp/:agentId", async (_, reply) => {
-    return reply.code(405).send(mcpMethodNotAllowed());
-  });
-
-  // --- Branding (public, no auth required) ---
-
-  app.get("/api/v1/app/branding", async () => {
-    return { iconColor: cachedIconColor };
-  });
-
-  // --- Release routes ---
-
-  app.get("/api/v1/app/version", async () => {
-    return getAppVersionInfo();
-  });
-
-  app.get("/api/v1/release/status", async () => {
-    const record = await readReleaseStore();
-    return { tag: record?.tag ?? null, deployedAt: record?.deployedAt ?? null };
-  });
-
-  app.get("/api/v1/release/info", async (_, reply) => {
-    try {
-      await runCommand("git", [
-        "-C",
-        serverDir,
-        "fetch",
-        "origin",
-        "--tags",
-        "--quiet",
-      ]);
-
-      // Current deployed tag from store (fast)
-      const record = await readReleaseStore();
-      const currentTag = record?.tag ?? null;
-
-      // Check admin permissions (cached for process lifetime)
-      const isAdmin = await checkIsAdmin();
-
-      // Read the configured release channel
-      const channelRaw = await getSetting(pool, "release_channel");
-      const channel = channelRaw === "latest" ? "latest" : "stable";
-
-      // Fetch recent releases in a single gh call. We request a small batch
-      // with isPrerelease so we can derive both the channel-filtered tag and
-      // the absolute latest tag without a second subprocess spawn.
-      let latestTag: string | null = null;
-      let absoluteLatestTag: string | null = null;
-      try {
-        const repo = await getGitHubRepo();
-        const ghResult = await runCommand("gh", [
-          "release",
-          "list",
-          "--repo",
-          repo,
-          "--limit",
-          "10",
-          "--json",
-          "tagName,isPrerelease",
-        ]);
-        const allReleases = parseGhJson<
-          Array<{ tagName: string; isPrerelease: boolean }>
-        >(ghResult.stdout);
-        absoluteLatestTag = allReleases[0]?.tagName ?? null;
-        if (channel === "stable") {
-          latestTag = allReleases.find((r) => !r.isPrerelease)?.tagName ?? null;
-        } else {
-          latestTag = allReleases[0]?.tagName ?? null;
-        }
-      } catch {
-        // Fallback to git tags (sees all releases regardless of channel)
-        const tagsResult = await runCommand("git", [
-          "-C",
-          serverDir,
-          "tag",
-          "--sort=-version:refname",
-        ]);
-        const fallbackTag =
-          tagsResult.stdout.split("\n").find((t) => t.startsWith("v")) ?? null;
-        latestTag = fallbackTag;
-        absoluteLatestTag = fallbackTag;
-      }
-
-      const updateAvailable = !!(
-        currentTag &&
-        latestTag &&
-        compareSemver(latestTag, currentTag) > 0
-      );
-
-      // Try to enrich with GitHub Release metadata
-      let latestRelease: {
-        tag: string;
-        publishedAt: string;
-        url: string;
-      } | null = null;
-      let assistedMetadata: AssistedUpdateMetadata | null = null;
-      let assistedMetadataError: string | null = null;
-      if (latestTag && updateAvailable) {
-        const fullRelease = await fetchLatestReleaseMetadata(latestTag);
-        latestRelease = fullRelease
-          ? {
-              tag: fullRelease.tag,
-              publishedAt: fullRelease.publishedAt,
-              url: fullRelease.url,
-            }
-          : null;
-        const inspected = inspectAssistedUpdateMetadata(
-          fullRelease?.body ?? null
-        );
-        if (inspected.state === "invalid") {
-          assistedMetadataError = inspected.error;
-        } else if (inspected.state === "valid") {
-          assistedMetadata = inspected.metadata;
-        }
-      }
-      if (assistedMetadataError) {
-        return reply.code(500).send({
-          error: `Latest release has malformed assisted-update metadata: ${assistedMetadataError}`,
-        });
-      }
-
-      // Pending migrations come from the target release tarball. Migration
-      // gating supersedes the legacy release-scoped `assisted` metadata
-      // (CRU-146) — `assistedRequired` is now derived from the count of
-      // unapplied migrations in the target tag. The legacy `assisted`
-      // field stays in the response only for transitional compatibility.
-      let pendingMigrations: PendingMigrationSummary[] = [];
-      let migrationsError: string | null = null;
-      if (latestTag && updateAvailable) {
-        try {
-          const repo = await getGitHubRepo();
-          const evaluation = await evaluatePendingMigrations(latestTag, {
-            repo,
-          });
-          pendingMigrations = evaluation.pending.map((m) =>
-            toSummary(m.manifest)
-          );
-          if (evaluation.errors.length > 0) {
-            migrationsError = evaluation.errors
-              .map((e) => `${e.filename}: ${e.error}`)
-              .join("; ");
-          }
-        } catch (err) {
-          migrationsError =
-            err instanceof Error ? err.message : "migration evaluation failed";
-        }
-      }
-      const assistedRequired =
-        pendingMigrations.length > 0 ||
-        // Fail closed when migration evaluation errored: we don't actually
-        // know whether the target ships migrations, so the UI must gate
-        // the one-click button. The /release/update endpoint will return
-        // 503 if invoked anyway. Pair this with `migrationsError` in the
-        // response so the UI can surface the underlying failure.
-        (migrationsError !== null && updateAvailable) ||
-        // Transitional fallback: a release authored with the legacy
-        // `dispatch-update` block but without migration manifests should
-        // still gate. Drop this branch once every gated release ships
-        // migrations instead of release-scoped metadata (issue step 8).
-        isAssistedUpdateRequired(assistedMetadata, currentTag);
-
-      // Admin-only: unreleased commits on main.
-      // Compare the absolute latest tag (across all channels) to main so
-      // admins see the real unreleased count regardless of channel.
-      let unreleasedCount = 0;
-      let commits: Array<{ sha: string; subject: string }> = [];
-      let refMissing = false;
-
-      const compareTag = absoluteLatestTag ?? currentTag;
-      if (isAdmin && compareTag) {
-        const refCheck = await runCommand(
-          "git",
-          ["-C", serverDir, "rev-parse", "--verify", compareTag],
-          { allowedExitCodes: [0, 128] }
-        );
-        if (refCheck.exitCode !== 0) {
-          refMissing = true;
-        } else {
-          const countResult = await runCommand("git", [
-            "-C",
-            serverDir,
-            "rev-list",
-            `${compareTag}..origin/main`,
-            "--count",
-          ]);
-          unreleasedCount = Number(countResult.stdout) || 0;
-
-          if (unreleasedCount > 0) {
-            const logResult = await runCommand("git", [
-              "-C",
-              serverDir,
-              "log",
-              `${compareTag}..origin/main`,
-              "--no-merges",
-              "--format=%H\t%s",
-              "--max-count=20",
-            ]);
-            commits = logResult.stdout
-              .split("\n")
-              .filter((l) => l.trim())
-              .map((line) => {
-                const tab = line.indexOf("\t");
-                return {
-                  sha: line.slice(0, tab).slice(0, 7),
-                  subject: line.slice(tab + 1),
-                };
-              });
-          }
-        }
-      }
-
-      return {
-        currentTag,
-        channel,
-        isAdmin,
-        latestTag,
-        updateAvailable,
-        latestRelease,
-        unreleasedCount,
-        commits,
-        refMissing,
-        assisted: assistedMetadata,
-        assistedRequired,
-        pendingMigrations,
-        migrationsError,
-      };
-    } catch (err) {
-      const message = err instanceof Error ? err.message : "Unknown error";
-      return reply.code(500).send({ error: message });
-    }
-  });
-
-  // --- Release channel ---
-
-  const RELEASE_CHANNEL_KEY = "release_channel";
-  const VALID_CHANNELS = ["stable", "latest"] as const;
-  type ReleaseChannel = (typeof VALID_CHANNELS)[number];
-
-  app.get("/api/v1/release/channel", async () => {
-    const raw = await getSetting(pool, RELEASE_CHANNEL_KEY);
-    const channel: ReleaseChannel = raw === "latest" ? "latest" : "stable";
-    return { channel };
-  });
-
-  app.post("/api/v1/release/channel", async (request, reply) => {
-    const body = request.body as { channel?: unknown } | undefined;
-    if (
-      !body?.channel ||
-      !VALID_CHANNELS.includes(body.channel as ReleaseChannel)
-    ) {
-      return reply.code(400).send({
-        error: `channel must be one of: ${VALID_CHANNELS.join(", ")}`,
-      });
-    }
-    await setSetting(pool, RELEASE_CHANNEL_KEY, body.channel as string);
-    return { channel: body.channel };
-  });
-
-  // --- Release admin check ---
-
-  app.get("/api/v1/release/admin-check", async () => {
-    const isAdmin = await checkIsAdmin();
-    return { isAdmin };
-  });
-
-  // --- Promote release (remove pre-release flag) ---
-
-  app.post("/api/v1/release/promote", async (request, reply) => {
-    const isAdmin = await checkIsAdmin();
-    if (!isAdmin) {
-      return reply.code(403).send({ error: "Admin access required" });
-    }
-
-    const body = request.body as { tag?: unknown } | undefined;
-    if (
-      !body?.tag ||
-      typeof body.tag !== "string" ||
-      !/^v\d+\.\d+\.\d+$/.test(body.tag)
-    ) {
-      return reply.code(400).send({
-        error: "tag is required and must be a semver tag (e.g. v0.11.18)",
-      });
-    }
-
-    try {
-      const repo = await getGitHubRepo();
-      await runCommand("gh", [
-        "release",
-        "edit",
-        body.tag,
-        "--repo",
-        repo,
-        "--prerelease=false",
-      ]);
-      return { ok: true, tag: body.tag };
-    } catch (err) {
-      const message = err instanceof Error ? err.message : "Unknown error";
-      return reply.code(500).send({ error: message });
-    }
-  });
-
-  // --- Recent releases list ---
-
-  app.get("/api/v1/releases", async (_, reply) => {
-    try {
-      const repo = await getGitHubRepo();
-      const result = await runCommand("gh", [
-        "release",
-        "list",
-        "--repo",
-        repo,
-        "--limit",
-        "10",
-        "--json",
-        "tagName,publishedAt,isPrerelease",
-      ]);
-      const releases = parseGhJson<
-        Array<{
-          tagName: string;
-          publishedAt: string;
-          isPrerelease: boolean;
-        }>
-      >(result.stdout);
-      return {
-        releases: releases.map((r) => ({
-          tag: r.tagName,
-          publishedAt: r.publishedAt,
-          isPrerelease: r.isPrerelease,
-          url: `https://github.com/${repo}/releases/tag/${r.tagName}`,
-        })),
-      };
-    } catch (err) {
-      const message = err instanceof Error ? err.message : "Unknown error";
-      return reply.code(500).send({ error: message });
-    }
-  });
-
-  app.post("/api/v1/release", async (request, reply) => {
-    const body = request.body as { versionType?: unknown } | undefined;
-
-    if (
-      !body?.versionType ||
-      !RELEASE_VERSION_TYPES.includes(body.versionType as ReleaseVersionType)
-    ) {
-      return reply.code(400).send({
-        error: `versionType must be one of: ${RELEASE_VERSION_TYPES.join(", ")}`,
-      });
-    }
-
-    const versionType = body.versionType as ReleaseVersionType;
-
-    // Only one release at a time
-    if (activeReleaseJob && !isTerminalPhase(activeReleaseJob.phase)) {
-      return reply
-        .code(409)
-        .send({ error: "A release is already in progress." });
-    }
-
-    // Quick pre-flight: check gh CLI before spawning the job
-    try {
-      await runCommand("gh", ["--version"]);
-    } catch {
-      return reply.code(422).send({
-        error:
-          "GitHub CLI (gh) is not available. Install it from https://cli.github.com",
-      });
-    }
-
-    const job: ReleaseJob = {
-      jobType: "create",
-      versionType,
-      phase: "preflight",
-      startedAt: new Date().toISOString(),
-      log: [],
-      runUrl: null,
-      tag: null,
-      error: null,
-    };
-    activeReleaseJob = job;
-
-    // Run async — do not await
-    void runReleaseJob(job);
-
-    return reply.code(202).send({ ok: true });
-  });
-
-  app.post("/api/v1/release/update", async (request, reply) => {
-    const body = request.body as { tag?: unknown } | undefined;
-    const bearerToken = getBearerToken(request);
-
-    if (
-      !body?.tag ||
-      typeof body.tag !== "string" ||
-      !/^v\d+\.\d+\.\d+$/.test(body.tag)
-    ) {
-      return reply.code(400).send({
-        error: "tag is required and must be a semver tag (e.g. v0.2.31)",
-      });
-    }
-
-    const tag = body.tag as string;
-    const releaseUpdateAgentId = bearerToken
-      ? getReleaseUpdateAgentId(config.authToken, bearerToken)
-      : null;
-
-    if (releaseUpdateAgentId) {
-      const agent = await agentManager.getAgent(releaseUpdateAgentId);
-      if (
-        !agent ||
-        agent.role !== "assisted_update" ||
-        agent.cwd !== serverDir
-      ) {
-        return reply
-          .code(403)
-          .send({ error: "Invalid assisted update token." });
-      }
-      // Bind the bypass token to the tag the agent was launched for. Without
-      // this, a holder of the bearer token could deploy any tag (skipping
-      // the migration gate at line 2611) once the assisted job reaches a
-      // terminal phase, because the takeover check below only fires for
-      // active non-terminal jobs. Matching against `assisted-update.json`
-      // keeps the bypass scope honest: one agent, one tag.
-      const assistedState = await readAssistedUpdateState().catch(() => null);
-      if (assistedState && assistedState.tag !== tag) {
-        return reply.code(403).send({
-          error:
-            "Assisted update token is bound to a different tag. Launch a fresh assisted update for this tag.",
-        });
-      }
-    }
-
-    // Only one release/update at a time, with one carve-out: the assisted
-    // agent's `apply` step posts to this endpoint, and at that point the
-    // active job is the agent's own update-assisted job. Without this
-    // exception, the agent is rejected by its own job and the documented
-    // path in the assisted prompt is unreachable. The on-disk assisted
-    // state remains the canonical phase tracker — replacing
-    // activeReleaseJob with the new update job is just for log/SSE
-    // routing during the deploy.
-    if (activeReleaseJob && !isTerminalPhase(activeReleaseJob.phase)) {
-      const isAssistedTakeover =
-        releaseUpdateAgentId !== null &&
-        activeReleaseJob.jobType === "update-assisted" &&
-        activeReleaseJob.assisted.agentId === releaseUpdateAgentId &&
-        activeReleaseJob.tag === tag;
-      if (!isAssistedTakeover) {
-        return reply
-          .code(409)
-          .send({ error: "A release or update is already in progress." });
-      }
-    }
-
-    // Gate: if the target release ships unapplied migration manifests, or
-    // (transitionally) declares `mode: required` via legacy metadata, the
-    // generic one-click path is not allowed. The UI must route the
-    // operator through the assisted-update agent launch instead. The
-    // launched agent itself bypasses this gate using its bearer token
-    // (releaseUpdateAgentId), so the recovery path still works once the
-    // assisted flow is in progress.
-    if (!releaseUpdateAgentId) {
-      const installed = await readReleaseStore().catch(() => null);
-
-      // Fail closed when migrations can't be evaluated. If we can't read
-      // `update-migrations/*` from the target tarball we don't actually
-      // know whether the target ships migrations — falling through to
-      // the legacy gate would let a transient network error bypass a
-      // mandatory assisted update once the legacy fence is removed
-      // (issue step 8). 503 is the right shape: the operator can retry
-      // when the underlying issue (download, extraction, parse) clears.
-      let pendingMigrationsForGate: PendingMigrationSummary[];
-      try {
-        const repo = await getGitHubRepo();
-        const evaluation = await evaluatePendingMigrations(tag, { repo });
-        pendingMigrationsForGate = evaluation.pending.map((m) =>
-          toSummary(m.manifest)
-        );
-      } catch (err) {
-        const message = err instanceof Error ? err.message : String(err);
-        console.warn(
-          `[release/update] migration evaluation failed for ${tag}: ${message}`
-        );
-        return reply.code(503).send({
-          error: "MIGRATION_EVALUATION_UNAVAILABLE",
-          message: `Could not evaluate update migrations for ${tag}: ${message}. Retry once the underlying issue clears.`,
-        });
-      }
-      if (pendingMigrationsForGate.length > 0) {
-        return reply.code(409).send({
-          error: "ASSISTED_UPDATE_REQUIRED",
-          message:
-            "This release has unapplied migrations. POST to /api/v1/release/assisted/launch instead.",
-          pendingMigrations: pendingMigrationsForGate,
-        });
-      }
-
-      const targetMeta = await fetchReleaseMetadata(tag);
-      const inspected = inspectAssistedUpdateMetadata(targetMeta?.body ?? null);
-      if (inspected.state === "invalid") {
-        return reply.code(409).send({
-          error: "ASSISTED_UPDATE_METADATA_INVALID",
-          message: `Target release has malformed assisted-update metadata: ${inspected.error}`,
-        });
-      }
-      const targetAssisted =
-        inspected.state === "valid" ? inspected.metadata : null;
-      if (isAssistedUpdateRequired(targetAssisted, installed?.tag ?? null)) {
-        return reply.code(409).send({
-          error: "ASSISTED_UPDATE_REQUIRED",
-          message:
-            "This release requires the assisted update flow. POST to /api/v1/release/assisted/launch instead.",
-          assisted: targetAssisted,
-        });
-      }
-    }
-
-    const job: ReleaseJob = {
-      jobType: "update",
-      versionType: null,
-      phase: "fetching",
-      startedAt: new Date().toISOString(),
-      log: [],
-      runUrl: null,
-      tag,
-      error: null,
-    };
-    activeReleaseJob = job;
-
-    // Run async — do not await
-    void runUpdateJob(job);
-
-    return reply.code(202).send({ ok: true });
-  });
-
-  app.post("/api/v1/release/assisted/launch", async (request, reply) => {
-    const body = request.body as { tag?: unknown } | undefined;
-
-    if (
-      !body?.tag ||
-      typeof body.tag !== "string" ||
-      !/^v\d+\.\d+\.\d+$/.test(body.tag)
-    ) {
-      return reply.code(400).send({
-        error: "tag is required and must be a semver tag (e.g. v0.2.31)",
-      });
-    }
-
-    const enabledAgentTypes = await getEnabledAgentTypes(pool);
-    const assistedType = enabledAgentTypes.find(isCliAgentType);
-    if (!assistedType) {
-      return reply.code(422).send({
-        error:
-          "No CLI agent types are enabled. Enable Codex, Claude, or OpenCode first.",
-      });
-    }
-
-    if (activeReleaseJob && !isTerminalPhase(activeReleaseJob.phase)) {
-      return reply
-        .code(409)
-        .send({ error: "A release or update is already in progress." });
-    }
-
-    if (activeAssistedUpdateLaunch) {
-      return reply.code(409).send({
-        error:
-          "An assisted update agent is already being created for the production checkout.",
-      });
-    }
-
-    activeAssistedUpdateLaunch = true;
-    try {
-      if (await hasActiveAssistedUpdateAgent()) {
-        return reply.code(409).send({
-          error:
-            "An assisted update agent is already active for the production checkout.",
-        });
-      }
-
-      const record = await readReleaseStore().catch(() => null);
-      const worktreeLocationRaw = await getSetting(pool, WORKTREE_LOCATION_KEY);
-      const worktreeLocation: WorktreeLocation =
-        worktreeLocationRaw &&
-        (VALID_WORKTREE_LOCATIONS as string[]).includes(worktreeLocationRaw)
-          ? (worktreeLocationRaw as WorktreeLocation)
-          : "sibling";
-
-      // Prefer the migration-driven path (CRU-146): load every pending
-      // migration manifest from the target tarball and pass them to the
-      // launched agent as one ordered plan. Fall back to the legacy
-      // release-scoped `dispatch-update` block if no migrations are
-      // pending — both paths land in the same `assisted-update.json`
-      // state and the same phase machinery.
-      let pendingMigrationManifests: UpdateMigrationManifest[] = [];
-      try {
-        const repo = await getGitHubRepo();
-        const evaluation = await evaluatePendingMigrations(body.tag, { repo });
-        if (evaluation.errors.length > 0) {
-          return reply.code(409).send({
-            error: "ASSISTED_UPDATE_MIGRATIONS_INVALID",
-            message: `Target release has malformed migration manifests: ${evaluation.errors
-              .map((e) => `${e.filename}: ${e.error}`)
-              .join("; ")}`,
-          });
-        }
-        pendingMigrationManifests = evaluation.pending.map((m) => m.manifest);
-      } catch (err) {
-        // Evaluator failures (download error, missing artifact for an
-        // older release, etc.) shouldn't kill the launch — fall through
-        // to the legacy `dispatch-update` body fence so older gated
-        // releases keep working.
-        const message = err instanceof Error ? err.message : String(err);
-        console.warn(
-          `[release/assisted/launch] migration evaluation failed for ${body.tag}, falling back to legacy metadata: ${message}`
-        );
-      }
-
-      const targetMeta = await fetchReleaseMetadata(body.tag);
-      const inspected = inspectAssistedUpdateMetadata(targetMeta?.body ?? null);
-      if (
-        pendingMigrationManifests.length === 0 &&
-        inspected.state === "invalid"
-      ) {
-        return reply.code(409).send({
-          error: "ASSISTED_UPDATE_METADATA_INVALID",
-          message: `Target release has malformed assisted-update metadata: ${inspected.error}`,
-        });
-      }
-      const assistedMeta =
-        inspected.state === "valid" ? inspected.metadata : null;
-      let assistedState: AssistedUpdateState | null = null;
-      let assistedToken: string | null = null;
-      let initialPrompt: string;
-      if (pendingMigrationManifests.length > 0 || assistedMeta) {
-        // The launched agent runs on the same host as the server, so we
-        // build the phase-callback URL from server config rather than the
-        // inbound request's Host header (which is attacker-controllable
-        // for non-browser clients). The framework prompt subsumes the
-        // recovery skeleton — it's the canonical instruction set when a
-        // release declares migrations or legacy metadata.
-        const baseUrl = dispatchBaseUrl();
-        const ctx = await buildAssistedUpdateContext(
-          {
-            tag: body.tag,
-            fromTag: record?.tag ?? null,
-            migrations:
-              pendingMigrationManifests.length > 0
-                ? pendingMigrationManifests
-                : undefined,
-            metadata:
-              pendingMigrationManifests.length === 0 && assistedMeta
-                ? assistedMeta
-                : undefined,
-            serverDir,
-            recovery: {
-              serviceCommand: defaultServiceRestartCommand(),
-              healthEndpoint: dispatchHealthUrl(),
-              serviceLogPath: "~/.dispatch/logs/dispatch.log",
-              failureLogPath: "~/.dispatch/logs/last-release-failure.log",
-            },
-          },
-          baseUrl
-        );
-        assistedState = ctx.state;
-        assistedToken = ctx.state.token;
-        initialPrompt = ctx.prompt;
-      } else {
-        // No migrations and no legacy metadata — fall back to the legacy
-        // recovery skeleton. This keeps the existing manual-rescue flow
-        // for releases that don't opt into the framework.
-        initialPrompt = buildAssistedUpdatePrompt({
-          tag: body.tag,
-          currentTag: record?.tag ?? null,
-        });
-      }
-
-      const agent = await agentManager.createAgent({
-        name: `update-${body.tag}`,
-        type: assistedType,
-        role: "assisted_update",
-        cwd: serverDir,
-        fullAccess: true,
-        useWorktree: false,
-        worktreeLocation,
-        initialPrompt,
-      });
-
-      if (assistedState && assistedToken) {
-        // First durable persistence of the assisted-update state. We
-        // intentionally don't write before createAgent succeeds — a
-        // half-completed launch would otherwise leave a phantom record
-        // on disk that rehydrateActiveAssistedJob() would resurrect on
-        // the next reconnect or process restart.
-        await attachAssistedAgent(assistedState, agent.id);
-        const launchLog = [
-          `==> assisted update launched for ${body.tag}`,
-          `==> agent: ${agent.id}`,
-        ];
-        if (pendingMigrationManifests.length > 0) {
-          launchLog.push(
-            `==> migrations: ${pendingMigrationManifests
-              .map((m) => m.id)
-              .join(", ")}`
-          );
-        } else if (assistedMeta) {
-          launchLog.push(`==> mode: ${assistedMeta.mode}`);
-        }
-        const job: ReleaseJob = {
-          jobType: "update-assisted",
-          versionType: null,
-          phase: "inspect",
-          startedAt: new Date().toISOString(),
-          log: launchLog,
-          runUrl: null,
-          tag: body.tag,
-          error: null,
-          assisted: { ...assistedState, agentId: agent.id },
-        };
-        activeReleaseJob = job;
-        broadcastReleaseEvent({
-          type: "assisted",
-          state: { ...assistedState, agentId: agent.id },
-        });
-      }
-
-      queueGitContextRefresh([agent.id]);
-      uiEventBroker.publish({
-        type: "agent.upsert",
-        agent: withStreamFlag(agent),
-      });
-      return reply
-        .code(201)
-        .send({ agent: withStreamFlag(agent), assisted: assistedState });
-    } catch (error) {
-      return handleAgentError(reply, error);
-    } finally {
-      activeAssistedUpdateLaunch = false;
-    }
-  });
-
-  app.post("/api/v1/release/assisted/phase", async (request, reply) => {
-    const body = request.body as
-      | { token?: unknown; phase?: unknown; note?: unknown; error?: unknown }
-      | undefined;
-    if (!body?.token || typeof body.token !== "string") {
-      return reply.code(400).send({ error: "token is required" });
-    }
-    if (!body.phase || typeof body.phase !== "string") {
-      return reply.code(400).send({ error: "phase is required" });
-    }
-    const result = await applyAssistedPhase({
-      token: body.token,
-      phase: body.phase as AssistedPhase,
-      note: typeof body.note === "string" ? body.note : undefined,
-      error: typeof body.error === "string" ? body.error : undefined,
-    });
-    if (!result.ok) {
-      return reply.code(409).send({ error: result.reason });
-    }
-    if (activeReleaseJob && activeReleaseJob.jobType === "update-assisted") {
-      activeReleaseJob.phase = result.state.phase;
-      activeReleaseJob.assisted = result.state;
-      if (result.state.error) activeReleaseJob.error = result.state.error;
-      // The persisted note is already sanitized (sanitizeAgentString in
-      // applyAssistedPhase clamps + strips newlines), so we can drop it
-      // straight into the log line.
-      const persistedNote = result.state.notes[result.state.phase];
-      const summary = persistedNote
-        ? `==> phase ${result.state.phase}: ${persistedNote}`
-        : `==> phase ${result.state.phase}`;
-      appendReleaseLog(activeReleaseJob, summary);
-      broadcastReleaseEvent({ type: "phase", phase: activeReleaseJob.phase });
-      broadcastReleaseEvent({ type: "assisted", state: result.state });
-    }
-
-    // When the agent reports `validate`, run the metadata-declared checks
-    // and gate the success transition. Failed checks move state to
-    // `blocked` so the operator UI shows exactly what gated the release.
-    if (result.state.phase === "validate") {
-      const post = await runAndRecordChecks(result.state, {
-        serverDir,
-        targetTag: result.state.tag,
-        healthUrl: dispatchHealthUrl(),
-      });
-      if (activeReleaseJob && activeReleaseJob.jobType === "update-assisted") {
-        activeReleaseJob.phase = post.phase;
-        activeReleaseJob.assisted = post;
-        if (post.error) activeReleaseJob.error = post.error;
-        for (const c of post.checks) {
-          appendReleaseLog(
-            activeReleaseJob,
-            `  - ${c.ok ? "✓" : "✗"} ${c.name}: ${c.message}`
-          );
-        }
-        broadcastReleaseEvent({ type: "phase", phase: activeReleaseJob.phase });
-        broadcastReleaseEvent({ type: "assisted", state: post });
-      }
-    }
-
-    return reply.code(200).send({ ok: true, state: result.state });
-  });
-
-  app.get("/api/v1/release/assisted/state", async () => {
-    const state = await readAssistedUpdateState();
-    return { state };
-  });
-
-  app.delete("/api/v1/release/assisted/state", async () => {
-    await clearAssistedUpdateState();
-    if (activeReleaseJob?.jobType === "update-assisted") {
-      activeReleaseJob = null;
-    }
-    return { ok: true };
-  });
-
-  app.get("/api/v1/release/stream", async (_request, reply) => {
-    reply.raw.setHeader("Content-Type", "text/event-stream");
-    reply.raw.setHeader("Cache-Control", "no-cache, no-transform");
-    reply.raw.setHeader("Connection", "keep-alive");
-    reply.raw.setHeader("X-Accel-Buffering", "no");
-    reply.hijack();
-
-    const stream = reply.raw;
-    releaseStreamClients.add(stream);
-
-    const heartbeat = setInterval(() => {
-      stream.write(": keepalive\n\n");
-    }, 20_000);
-
-    // If the server was restarted mid-assisted-update, in-memory state
-    // is gone but the on-disk file still has the reached phase. Pull
-    // it back into `activeReleaseJob` before snapshotting so the UI
-    // sees a real in-flight job rather than null.
-    await rehydrateActiveAssistedJob();
-
-    // Send current job snapshot. For an in-flight assisted update, refresh
-    // from disk so a late reconnect rehydrates phase + checks even if the
-    // agent crashed mid-run between events.
-    let snapshotJob = activeReleaseJob;
-    if (snapshotJob && snapshotJob.jobType === "update-assisted") {
-      const persisted = await readAssistedUpdateState();
-      if (persisted) {
-        snapshotJob = { ...snapshotJob, assisted: persisted };
-      }
-    }
-    const snapshot: ReleaseStreamEvent = {
-      type: "snapshot",
-      job: snapshotJob,
-    };
-    stream.write(`data: ${JSON.stringify(snapshot)}\n\n`);
-
-    stream.on("close", () => {
-      clearInterval(heartbeat);
-      releaseStreamClients.delete(stream);
-    });
-  });
-
-  // --- Health / existing routes ---
-
-  app.get("/api/v1/health", async () => {
-    const result = await pool.query("SELECT NOW() AS now");
-
-    return {
-      status: "ok",
-      db: "ok",
-      now: result.rows[0]?.now,
-    };
-  });
-
-  app.get("/ping", async () => {
-    return { status: "ok" };
-  });
-
-  // Write an image from the browser clipboard to the host's system clipboard.
-  // This bridges remote browser sessions to the local system clipboard so that
-  // CLI tools (e.g. Claude CLI) can read pasted images via native APIs.
-  app.post("/api/v1/clipboard/image", async (request, reply) => {
-    const data = await request.file();
-    if (!data) {
-      return reply
-        .code(400)
-        .send({ error: "An image file field is required." });
-    }
-    const mime = data.mimetype;
-    if (!mime.startsWith("image/")) {
-      return reply.code(400).send({ error: "Only image files are accepted." });
-    }
-
-    const buffer = await data.toBuffer();
-    const ext =
-      mime === "image/png" ? "png" : mime === "image/jpeg" ? "jpg" : "png";
-    const tmpPath = `/tmp/dispatch-clipboard-${Date.now()}.${ext}`;
-    await writeFile(tmpPath, buffer);
-
-    try {
-      if (os.platform() === "darwin") {
-        const pasteboardClass = ext === "jpg" ? "JPEG" : "PNGf";
-        await new Promise<void>((resolve, reject) => {
-          const proc = spawn("osascript", [
-            "-e",
-            `set the clipboard to (read (POSIX file "${tmpPath}") as «class ${pasteboardClass}»)`,
-          ]);
-          proc.on("close", (code) =>
-            code === 0
-              ? resolve()
-              : reject(new Error(`osascript exited ${code}`))
-          );
-          proc.on("error", reject);
-        });
-      } else {
-        const display = process.env.DISPATCH_COPY_DISPLAY;
-        if (!display) {
-          return reply.code(500).send({
-            error:
-              "DISPATCH_COPY_DISPLAY is not set. Clipboard image paste on Linux requires Xvfb and xclip.",
-          });
-        }
-        await new Promise<void>((resolve, reject) => {
-          const proc = spawn(
-            "xclip",
-            ["-selection", "clipboard", "-t", mime, "-i", tmpPath],
-            {
-              env: { ...process.env, DISPLAY: display },
-            }
-          );
-          proc.on("close", (code) =>
-            code === 0 ? resolve() : reject(new Error(`xclip exited ${code}`))
-          );
-          proc.on("error", reject);
-        });
-      }
-      return reply.code(200).send({ ok: true });
-    } catch (err) {
-      const message = err instanceof Error ? err.message : String(err);
-      return reply
-        .code(500)
-        .send({ error: `Failed to write to clipboard: ${message}` });
-    } finally {
-      await unlink(tmpPath).catch(() => {});
-    }
-  });
-
-  app.get("/api/v1/system/defaults", async () => {
-    return {
-      homeDir: os.homedir(),
-    };
-  });
-
-  function resolveTildePath(raw: string): string {
-    if (raw.startsWith("~/")) return path.join(os.homedir(), raw.slice(2));
-    if (raw === "~") return os.homedir();
-    return raw;
-  }
-
-  app.get("/api/v1/system/path-info", async (request, reply) => {
-    const query = request.query as { path?: unknown };
-    if (typeof query?.path !== "string" || !query.path.trim()) {
-      return reply
-        .code(400)
-        .send({ error: "path query parameter is required." });
-    }
-    const resolved = resolveTildePath(query.path.trim());
-    if (!path.isAbsolute(resolved)) {
-      return {
-        exists: false,
-        isDirectory: false,
-        isGitRepo: false,
-        privacyRestricted: false,
-        resolvedPath: resolved,
-      };
-    }
-    if (shouldSkipAutomaticMacPathProbe(resolved, os.homedir())) {
-      return {
-        exists: false,
-        isDirectory: false,
-        isGitRepo: false,
-        privacyRestricted: true,
-        resolvedPath: resolved,
-      };
-    }
-    try {
-      const info = await stat(resolved).catch(() => null);
-      const exists = info !== null;
-      const isDirectory = exists && info.isDirectory();
-      let isGitRepo = false;
-      if (isDirectory) {
-        const result = await runCommand(
-          "git",
-          ["-C", resolved, "rev-parse", "--is-inside-work-tree"],
-          {
-            timeoutMs: 3_000,
-            allowedExitCodes: [0, 1, 128],
-          }
-        );
-        isGitRepo = result.exitCode === 0 && result.stdout.trim() === "true";
-      }
-      return {
-        exists,
-        isDirectory,
-        isGitRepo,
-        privacyRestricted: false,
-        resolvedPath: resolved,
-      };
-    } catch {
-      return {
-        exists: false,
-        isDirectory: false,
-        isGitRepo: false,
-        privacyRestricted: false,
-        resolvedPath: resolved,
-      };
-    }
-  });
-
-  app.get("/api/v1/system/path-completions", async (request, reply) => {
-    const query = request.query as { prefix?: unknown };
-    if (typeof query?.prefix !== "string" || !query.prefix.trim()) {
-      return reply
-        .code(400)
-        .send({ error: "prefix query parameter is required." });
-    }
-    const raw = query.prefix.trim();
-    const resolved = resolveTildePath(raw);
-    if (!path.isAbsolute(resolved)) {
-      return { completions: [] };
-    }
-    try {
-      const parentDir = path.dirname(resolved);
-      const partial = path.basename(resolved).toLowerCase();
-
-      // If the prefix ends with /, list children of that directory instead
-      const isExactDir = raw.endsWith("/");
-      const searchDir = isExactDir ? resolved : parentDir;
-      const searchPartial = isExactDir ? "" : partial;
-
-      if (shouldSkipAutomaticMacPathProbe(searchDir, os.homedir())) {
-        return { completions: [], privacyRestricted: true };
-      }
-
-      const entries = await readdir(searchDir, { withFileTypes: true });
-      const dirs = entries
-        .filter((entry) => {
-          if (!entry.isDirectory()) return false;
-          // Skip hidden dirs unless the partial starts with "."
-          if (entry.name.startsWith(".") && !searchPartial.startsWith("."))
-            return false;
-          return entry.name.toLowerCase().startsWith(searchPartial);
-        })
-        .sort((a, b) => a.name.localeCompare(b.name))
-        .slice(0, 20)
-        .map((entry) => path.join(searchDir, entry.name));
-
-      // Convert back to tilde paths if the input used tilde
-      const homeDir = os.homedir();
-      const completions = dirs.map((dir) =>
-        raw.startsWith("~") && dir.startsWith(homeDir)
-          ? "~" + dir.slice(homeDir.length)
-          : dir
-      );
-
-      return { completions, privacyRestricted: false };
-    } catch {
-      return { completions: [], privacyRestricted: false };
-    }
-  });
-
-  // --- Git helpers ---
-  app.get("/api/v1/git/branches", async (request, reply) => {
-    const query = request.query as { cwd?: unknown };
-    if (typeof query?.cwd !== "string" || !query.cwd.trim()) {
-      return reply
-        .code(400)
-        .send({ error: "cwd query parameter is required." });
-    }
-    try {
-      const rawCwd = query.cwd.trim();
-      const cwd = rawCwd.startsWith("~/")
-        ? path.join(os.homedir(), rawCwd.slice(2))
-        : rawCwd === "~"
-          ? os.homedir()
-          : rawCwd;
-      const result = await runCommand(
-        "git",
-        ["-C", cwd, "ls-remote", "--heads", "origin"],
-        {
-          timeoutMs: 15_000,
-        }
-      );
-      if (result.exitCode !== 0) {
-        return reply
-          .code(500)
-          .send({ error: "Failed to list remote branches." });
-      }
-      const branches = result.stdout
-        .split("\n")
-        .map((line) => line.replace(/^.*refs\/heads\//, "").trim())
-        .filter(Boolean)
-        .sort((a, b) => {
-          if (a === "main") return -1;
-          if (b === "main") return 1;
-          if (a === "master") return -1;
-          if (b === "master") return 1;
-          return a.localeCompare(b);
-        });
-      return { branches };
-    } catch {
-      return reply.code(500).send({ error: "Failed to list remote branches." });
-    }
-  });
-
-  // --- Agent worktree settings ---
-  const WORKTREE_LOCATION_KEY = "worktree_location";
-  const INSTANCE_NAME_KEY = "instance_name";
-  type WorktreeLocation = "sibling" | "nested";
-  const VALID_WORKTREE_LOCATIONS: WorktreeLocation[] = ["sibling", "nested"];
-
-  app.get("/api/v1/agents/settings", async () => {
-    const raw = await getSetting(pool, WORKTREE_LOCATION_KEY);
-    const worktreeLocation: WorktreeLocation =
-      raw && (VALID_WORKTREE_LOCATIONS as string[]).includes(raw)
-        ? (raw as WorktreeLocation)
-        : "sibling";
-    const instanceName = (await getSetting(pool, INSTANCE_NAME_KEY)) ?? "";
-    return { worktreeLocation, iconColor: cachedIconColor, instanceName };
-  });
-
-  app.post("/api/v1/agents/settings", async (request, reply) => {
-    const body = request.body as {
-      worktreeLocation?: unknown;
-      iconColor?: unknown;
-      instanceName?: unknown;
-    };
-
-    if (body.worktreeLocation !== undefined) {
-      if (
-        typeof body.worktreeLocation !== "string" ||
-        !(VALID_WORKTREE_LOCATIONS as string[]).includes(body.worktreeLocation)
-      ) {
-        return reply
-          .code(400)
-          .send({ error: `worktreeLocation must be "sibling" or "nested".` });
-      }
-      await setSetting(pool, WORKTREE_LOCATION_KEY, body.worktreeLocation);
-    }
-
-    if (body.iconColor !== undefined) {
-      if (
-        typeof body.iconColor !== "string" ||
-        !(VALID_ICON_COLORS as readonly string[]).includes(body.iconColor)
-      ) {
-        return reply.code(400).send({
-          error: `iconColor must be one of: ${VALID_ICON_COLORS.join(", ")}`,
-        });
-      }
-      await setSetting(pool, ICON_COLOR_KEY, body.iconColor);
-      rewriteForColor(body.iconColor as IconColor);
-    }
-
-    if (body.instanceName !== undefined) {
-      if (typeof body.instanceName !== "string") {
-        return reply
-          .code(400)
-          .send({ error: "instanceName must be a string." });
-      }
-      const trimmed = body.instanceName.trim().slice(0, 100);
-      if (trimmed) {
-        await setSetting(pool, INSTANCE_NAME_KEY, trimmed);
-      } else {
-        await deleteSetting(pool, INSTANCE_NAME_KEY);
-      }
-    }
-
-    const raw = await getSetting(pool, WORKTREE_LOCATION_KEY);
-    const worktreeLocation: WorktreeLocation =
-      raw && (VALID_WORKTREE_LOCATIONS as string[]).includes(raw)
-        ? (raw as WorktreeLocation)
-        : "sibling";
-    const instanceName = (await getSetting(pool, INSTANCE_NAME_KEY)) ?? "";
-    return { worktreeLocation, iconColor: cachedIconColor, instanceName };
-  });
-
-  // --- Notification settings ---
-  app.get("/api/v1/notifications/settings", async () => {
-    return slackNotifier.getSettings();
-  });
-
-  app.post("/api/v1/notifications/settings", async (request, reply) => {
-    const body = request.body as {
-      webhookUrl?: unknown;
-      notifyEvents?: unknown;
-      webNotifyEnabled?: unknown;
-      webNotifyEvents?: unknown;
-    } | null;
-
-    if (body?.webhookUrl !== undefined) {
-      if (typeof body.webhookUrl !== "string") {
-        return reply.code(400).send({ error: "webhookUrl must be a string." });
-      }
-      if (body.webhookUrl && !isValidSlackWebhookUrl(body.webhookUrl)) {
-        return reply.code(400).send({
-          error: "webhookUrl must start with https://hooks.slack.com/",
-        });
-      }
-      await slackNotifier.setWebhookUrl(body.webhookUrl);
-    }
-
-    if (body?.notifyEvents !== undefined) {
-      if (!Array.isArray(body.notifyEvents)) {
-        return reply
-          .code(400)
-          .send({ error: "notifyEvents must be an array." });
-      }
-      await slackNotifier.setNotifyEvents(body.notifyEvents as string[]);
-    }
-
-    if (body?.webNotifyEnabled !== undefined) {
-      if (typeof body.webNotifyEnabled !== "boolean") {
-        return reply
-          .code(400)
-          .send({ error: "webNotifyEnabled must be a boolean." });
-      }
-      await slackNotifier.setWebNotifyEnabled(body.webNotifyEnabled);
-    }
-
-    if (body?.webNotifyEvents !== undefined) {
-      if (!Array.isArray(body.webNotifyEvents)) {
-        return reply
-          .code(400)
-          .send({ error: "webNotifyEvents must be an array." });
-      }
-      await slackNotifier.setWebNotifyEvents(body.webNotifyEvents as string[]);
-    }
-
-    return slackNotifier.getSettings();
-  });
-
-  app.post("/api/v1/notifications/test", async (request, reply) => {
-    const body = request.body as { webhookUrl?: unknown } | null;
-    const url =
-      typeof body?.webhookUrl === "string"
-        ? body.webhookUrl
-        : await slackNotifier.getWebhookUrl();
-    if (!url) {
-      return reply
-        .code(400)
-        .send({ error: "No webhook URL provided or configured." });
-    }
-    if (!isValidSlackWebhookUrl(url)) {
-      return reply
-        .code(400)
-        .send({ error: "webhookUrl must start with https://hooks.slack.com/" });
-    }
-    return slackNotifier.sendTestMessage(url);
-  });
-
-  app.get("/api/v1/app/settings/agent-types", async () => {
-    return {
-      enabledAgentTypes: await getEnabledAgentTypes(pool),
-    };
-  });
-
-  app.post("/api/v1/app/settings/agent-types", async (request, reply) => {
-    const body = request.body as { enabledAgentTypes?: unknown } | null;
-
-    if (!Array.isArray(body?.enabledAgentTypes)) {
-      return reply
-        .code(400)
-        .send({ error: "enabledAgentTypes must be an array." });
-    }
-
-    const uniqueTypes = body.enabledAgentTypes
-      .filter(
-        (value): value is (typeof AGENT_TYPES)[number] =>
-          typeof value === "string" &&
-          AGENT_TYPES.includes(value as (typeof AGENT_TYPES)[number])
-      )
-      .filter((value, index, values) => values.indexOf(value) === index);
-
-    if (uniqueTypes.length === 0) {
-      return reply
-        .code(400)
-        .send({ error: "At least one agent type must remain enabled." });
-    }
-
-    if (uniqueTypes.length !== body.enabledAgentTypes.length) {
-      return reply.code(400).send({
-        error: `enabledAgentTypes must only include ${AGENT_TYPES.join(", ")}.`,
-      });
-    }
-
-    return {
-      enabledAgentTypes: await setEnabledAgentTypes(pool, uniqueTypes),
-    };
-  });
-
-  // --- Energy metrics beacon (PWA diagnostics) ---
-  app.post("/api/v1/energy-report", async (request, reply) => {
-    try {
-      const metrics = request.body;
-      app.log.info({ energyMetrics: metrics }, "PWA energy metrics report");
-    } catch {
-      // Beacon payloads are fire-and-forget; don't fail loudly
-    }
-    return reply.status(204).send();
-  });
-
-  app.get("/api/v1/diagnostics/git-context", async () => {
-    const now = Date.now();
-    const pendingAges = Array.from(pendingGitRefreshEnqueuedAt.values()).map(
-      (queuedAt) => Math.max(0, now - queuedAt)
-    );
-    const oldestPendingAgeMs =
-      pendingAges.length > 0 ? Math.max(...pendingAges) : null;
-
-    const durations = [...gitRefreshDurationsMs].sort((a, b) => a - b);
-    const p50DurationMs = percentile(durations, 0.5);
-    const p95DurationMs = percentile(durations, 0.95);
-    const maxDurationMs =
-      durations.length > 0 ? durations[durations.length - 1] : null;
-    const lastDurationMs =
-      gitRefreshDurationsMs.length > 0
-        ? gitRefreshDurationsMs[gitRefreshDurationsMs.length - 1]
-        : null;
-
-    const agents = Array.from(gitRefreshAgentDiagnostics.entries())
-      .map(([agentId, diag]) => ({
-        agentId,
-        pending: pendingGitRefreshAgentIds.has(agentId),
-        active: activeGitRefreshAgentIds.has(agentId),
-        lastQueuedAt: toIso(diag.lastQueuedAt),
-        lastStartedAt: toIso(diag.lastStartedAt),
-        lastCompletedAt: toIso(diag.lastCompletedAt),
-        lastDurationMs: diag.lastDurationMs,
-        lastResult: diag.lastResult,
-        lastError: diag.lastError,
-      }))
-      .sort((a, b) => a.agentId.localeCompare(b.agentId));
-
-    return {
-      config: {
-        intervalMs: GIT_CONTEXT_REFRESH_INTERVAL_MS,
-        concurrency: GIT_CONTEXT_REFRESH_CONCURRENCY,
-        probeTimeoutMs: PROBE_COMMAND_TIMEOUT_MS,
-      },
-      queue: {
-        pending: pendingGitRefreshAgentIds.size,
-        active: activeGitRefreshAgentIds.size,
-        oldestPendingAgeMs,
-      },
-      counters: gitRefreshCounters,
-      durationsMs: {
-        samples: durations.length,
-        p50: p50DurationMs,
-        p95: p95DurationMs,
-        max: maxDurationMs,
-        last: lastDurationMs,
-      },
-      agents,
-    };
-  });
-
-  // ── Activity tracking ──────────────────────────────────────────────
-
-  app.get("/api/v1/activity/heatmap", async (request) => {
-    const query = request.query as Record<string, unknown>;
-    const days = Math.min(
-      Math.max(parseInt((query.days as string) ?? "365", 10) || 365, 1),
-      730
-    );
-    const rawTz =
-      typeof query.tz === "string" && query.tz ? query.tz : FALLBACK_TZ;
-    const tz = VALID_TIMEZONES.has(rawTz) ? rawTz : FALLBACK_TZ;
-
-    const result = await pool.query<{ day: string; count: number }>(
-      `SELECT ${dateTruncTz("day", "created_at", tz)} AS day, COUNT(*)::int AS count
-       FROM agent_events
-       WHERE created_at >= NOW() - make_interval(days => $1)
-       GROUP BY day ORDER BY day`,
-      [days]
-    );
-
-    return { days: result.rows };
-  });
-
-  app.get("/api/v1/activity/stats", async (request) => {
-    const aq = parseActivityQuery(request.query as Record<string, unknown>);
-    const { rows, rangeStart } = await loadScopedActivityEvents(aq);
-    const eventFilter = timeRangeClause(aq, "created_at");
-
-    const busiestDayResult = await pool.query<{ day: string; count: number }>(
-      `SELECT ${dateTruncTz("day", "created_at", aq.tz)} AS day, COUNT(*)::int AS count
-       FROM agent_events
-       ${eventFilter.clause}
-      GROUP BY day ORDER BY count DESC LIMIT 1`,
-      eventFilter.params
-    );
-    const stats = computeActivityStats(rows, rangeStart);
-
-    return {
-      totalWorkingMs: stats.totalWorkingMs,
-      avgBlockedMs: stats.avgBlockedMs,
-      avgWaitingMs: stats.avgWaitingMs,
-      busiestDay: busiestDayResult.rows[0]?.day ?? null,
-      busiestDayCount: busiestDayResult.rows[0]?.count ?? 0,
-      stateDurations: stats.stateDurations,
-    };
-  });
-
-  app.get("/api/v1/activity/daily-status", async (request) => {
-    const aq = parseActivityQuery(request.query as Record<string, unknown>);
-    const { rows, rangeStart } = await loadScopedActivityEvents(aq);
-    const dailyStatus = computeDailyStatus(rows, rangeStart, aq.granularity);
-
-    return { days: dailyStatus, granularity: aq.granularity };
-  });
-
-  app.get("/api/v1/activity/active-hours", async (request) => {
-    const aq = parseActivityQuery(request.query as Record<string, unknown>);
-    const eventFilter = timeRangeClause(aq, "created_at");
-    const result = await pool.query<{ created_at: string }>(
-      `SELECT created_at::text AS created_at
-       FROM agent_events
-       ${eventFilter.clause ? `${eventFilter.clause} AND` : "WHERE"} event_type IN ('working', 'blocked', 'waiting_user')
-       ORDER BY created_at`,
-      eventFilter.params
-    );
-
-    return { events: result.rows };
-  });
-
-  app.get("/api/v1/activity/agents-created", async (request) => {
-    const aq = parseActivityQuery(request.query as Record<string, unknown>);
-    const eventFilter = timeRangeClause(aq, "first_seen");
-    const result = await pool.query<{ day: string; count: number }>(
-      `SELECT ${dateTruncTz(aq.granularity, "first_seen", aq.tz)} AS day, COUNT(*)::int AS count
-       FROM (
-         SELECT agent_id, MIN(created_at) AS first_seen
-         FROM agent_events
-         GROUP BY agent_id
-       ) per_agent
-       ${eventFilter.clause}
-       GROUP BY day ORDER BY day`,
-      eventFilter.params
-    );
-    const total = result.rows.reduce((sum, r) => sum + r.count, 0);
-    return { days: result.rows, total, granularity: aq.granularity };
-  });
-
-  app.get("/api/v1/activity/working-time-by-project", async (request) => {
-    const aq = parseActivityQuery(request.query as Record<string, unknown>);
-    const rangeStart = aq.start;
-    const eventFilter = timeRangeClause(aq, "ae.created_at");
-
-    const inRangeResult = await pool.query<ActivityEventRow>(
-      `SELECT ae.agent_id, ae.event_type, ae.created_at,
-              COALESCE(ae.project_dir, a.cwd) AS project_dir
-       FROM agent_events ae
-       LEFT JOIN agents a ON a.id = ae.agent_id
-       ${eventFilter.clause}
-       ORDER BY ae.agent_id, ae.created_at`,
-      eventFilter.params
-    );
-
-    let rows = inRangeResult.rows;
-    if (rangeStart) {
-      const boundaryResult = await pool.query<ActivityEventRow>(
-        `SELECT DISTINCT ON (ae.agent_id) ae.agent_id, ae.event_type, ae.created_at,
-                COALESCE(ae.project_dir, a.cwd) AS project_dir
-         FROM agent_events ae
-         LEFT JOIN agents a ON a.id = ae.agent_id
-         WHERE ae.created_at < $1
-         ORDER BY ae.agent_id, ae.created_at DESC`,
-        [rangeStart]
-      );
-      rows = [...boundaryResult.rows, ...inRangeResult.rows].sort((a, b) => {
-        const agentCompare = a.agent_id.localeCompare(b.agent_id);
-        if (agentCompare !== 0) return agentCompare;
-        return a.created_at.getTime() - b.created_at.getTime();
-      });
-    }
-
-    const projects = computeWorkingTimeByProject(rows, rangeStart);
-    return { projects };
-  });
-
-  // ── Token usage ──────────────────────────────────────────────────
-
-  app.get("/api/v1/activity/token-stats", async (request) => {
-    const aq = parseActivityQuery(request.query as Record<string, unknown>);
-    const tokenFilter = timeRangeClause(
-      aq,
-      "COALESCE(session_start, harvested_at)"
-    );
-    const result = await pool.query<{
-      total_input: number;
-      total_cache_creation: number;
-      total_cache_read: number;
-      total_output: number;
-      total_messages: number;
-      total_sessions: number;
-    }>(
-      `SELECT
-        COALESCE(SUM(input_tokens), 0) AS total_input,
-        COALESCE(SUM(cache_creation_tokens), 0) AS total_cache_creation,
-        COALESCE(SUM(cache_read_tokens), 0) AS total_cache_read,
-        COALESCE(SUM(output_tokens), 0) AS total_output,
-        COALESCE(SUM(message_count), 0) AS total_messages,
-        COUNT(DISTINCT session_id) AS total_sessions
-       FROM agent_token_usage
-       ${tokenFilter.clause}`,
-      tokenFilter.params
-    );
-    return (
-      result.rows[0] ?? {
-        total_input: 0,
-        total_cache_creation: 0,
-        total_cache_read: 0,
-        total_output: 0,
-        total_messages: 0,
-        total_sessions: 0,
-      }
-    );
-  });
-
-  app.get("/api/v1/activity/token-daily", async (request) => {
-    const aq = parseActivityQuery(request.query as Record<string, unknown>);
-    const tokenFilter = timeRangeClause(
-      aq,
-      "COALESCE(session_start, harvested_at)"
-    );
-
-    const result = await pool.query<{
-      day: string;
-      input_tokens: number;
-      cache_creation_tokens: number;
-      cache_read_tokens: number;
-      output_tokens: number;
-      messages: number;
-    }>(
-      `SELECT
-        ${dateTruncTz(aq.granularity, "COALESCE(session_start, harvested_at)", aq.tz)} AS day,
-        SUM(input_tokens) AS input_tokens,
-        SUM(cache_creation_tokens) AS cache_creation_tokens,
-        SUM(cache_read_tokens) AS cache_read_tokens,
-        SUM(output_tokens) AS output_tokens,
-        SUM(message_count) AS messages
-       FROM agent_token_usage
-       ${tokenFilter.clause}
-       GROUP BY day ORDER BY day`,
-      tokenFilter.params
-    );
-    return { days: result.rows, granularity: aq.granularity };
-  });
-
-  app.get("/api/v1/activity/token-by-project", async (request) => {
-    const aq = parseActivityQuery(request.query as Record<string, unknown>);
-    const tokenFilter = timeRangeClause(
-      aq,
-      "COALESCE(t.session_start, t.harvested_at)"
-    );
-    const result = await pool.query<{
-      project_dir: string;
-      total_input: number;
-      total_output: number;
-      messages: number;
-    }>(
-      `SELECT
-        COALESCE(a.git_context->>'repoRoot', a.cwd) AS project_dir,
-        SUM(t.input_tokens + t.cache_creation_tokens + t.cache_read_tokens) AS total_input,
-        SUM(t.output_tokens) AS total_output,
-        SUM(t.message_count) AS messages
-       FROM agent_token_usage t
-       JOIN agents a ON a.id = t.agent_id
-       ${tokenFilter.clause}
-       GROUP BY project_dir
-       ORDER BY total_input DESC
-       LIMIT 20`,
-      tokenFilter.params
-    );
-    return { projects: result.rows };
-  });
-
-  app.get("/api/v1/activity/token-by-model", async (request) => {
-    const aq = parseActivityQuery(request.query as Record<string, unknown>);
-    const tokenFilter = timeRangeClause(
-      aq,
-      "COALESCE(session_start, harvested_at)"
-    );
-    const result = await pool.query<{
-      model: string;
-      total_input: number;
-      total_cache_creation: number;
-      total_cache_read: number;
-      total_output: number;
-      sessions: number;
-    }>(
-      `SELECT
-        model,
-        COALESCE(SUM(input_tokens), 0) AS total_input,
-        COALESCE(SUM(cache_creation_tokens), 0) AS total_cache_creation,
-        COALESCE(SUM(cache_read_tokens), 0) AS total_cache_read,
-        COALESCE(SUM(output_tokens), 0) AS total_output,
-       COUNT(DISTINCT session_id) AS sessions
-       FROM agent_token_usage
-       ${tokenFilter.clause}
-       GROUP BY model
-       ORDER BY (SUM(input_tokens) + SUM(cache_creation_tokens) + SUM(cache_read_tokens) + SUM(output_tokens)) DESC`,
-      tokenFilter.params
-    );
-    return { models: result.rows };
-  });
-
-  app.post("/api/v1/agents/:id/harvest-tokens", async (request, reply) => {
-    const { id } = request.params as { id: string };
-    const agent = await agentManager.getAgent(id);
-    if (!agent) return reply.code(404).send({ error: "Agent not found" });
-
-    await agentManager.harvestAgentTokens(agent);
-
-    return { ok: true };
-  });
-
-  // ── Agent History ─────────────────────────────────────────────────
-
-  app.get("/api/v1/history/projects", async () => {
-    const result = await pool.query<{ project: string }>(
-      `SELECT DISTINCT COALESCE(git_context->>'repoRoot', cwd) AS project
-       FROM agents
-       WHERE COALESCE(git_context->>'repoRoot', cwd) IS NOT NULL
-       ORDER BY project`
-    );
-    return { projects: result.rows.map((r) => r.project) };
-  });
-
-  app.get("/api/v1/history/agents", async (request) => {
-    const query = request.query as Record<string, unknown>;
-    const aq = parseActivityQuery(query);
-    const limit = Math.min(
-      Math.max(parseInt(String(query.limit ?? "50"), 10) || 50, 1),
-      100
-    );
-    const offset = Math.max(parseInt(String(query.offset ?? "0"), 10) || 0, 0);
-    const search = typeof query.search === "string" ? query.search.trim() : "";
-    const type = typeof query.type === "string" ? query.type : "";
-    const project = typeof query.project === "string" ? query.project : "";
-    const sortCol =
-      typeof query.sort === "string" &&
-      ["created_at", "name", "updated_at"].includes(query.sort)
-        ? query.sort
-        : "created_at";
-    const order =
-      typeof query.order === "string" && query.order === "asc" ? "ASC" : "DESC";
-
-    const conditions: string[] = ["a.parent_agent_id IS NULL"];
-    const params: unknown[] = [];
-
-    if (search) {
-      params.push(`%${escapeLike(search)}%`);
-      conditions.push(`a.name ILIKE $${params.length}`);
-    }
-    if (type) {
-      params.push(type);
-      conditions.push(`a.type = $${params.length}`);
-    }
-    if (project) {
-      params.push(project);
-      conditions.push(
-        `COALESCE(a.git_context->>'repoRoot', a.cwd) = $${params.length}`
-      );
-    }
-
-    const dateRange = timeRangeClause(aq, "a.created_at", params.length);
-    params.push(...dateRange.params);
-    if (dateRange.params.length > 0) {
-      // Extract conditions from the WHERE clause generated by timeRangeClause
-      const rangeConditions = dateRange.clause.replace(/^WHERE\s+/i, "");
-      conditions.push(rangeConditions);
-    }
-
-    const whereClause =
-      conditions.length > 0 ? `WHERE ${conditions.join(" AND ")}` : "";
-    const sortSql =
-      sortCol === "name"
-        ? `a.name ${order}, a.created_at DESC`
-        : `a.${sortCol} ${order}`;
-    const listParams = [...params, limit, offset];
-
-    const [countResult, agentsResult] = await Promise.all([
-      pool.query<{ total: number }>(
-        `SELECT COUNT(*)::int AS total FROM agents a ${whereClause}`,
-        params
-      ),
-      pool.query(
-        `SELECT
-          a.id,
-          a.name,
-          a.type,
-          a.status,
-          a.cwd,
-          a.worktree_path AS "worktreePath",
-          a.worktree_branch AS "worktreeBranch",
-          CASE
-            WHEN a.latest_event_type IS NULL OR a.latest_event_message IS NULL OR a.latest_event_updated_at IS NULL THEN NULL
-            ELSE json_build_object(
-              'type', a.latest_event_type,
-              'message', a.latest_event_message,
-              'updatedAt', a.latest_event_updated_at,
-              'metadata', COALESCE(a.latest_event_metadata, '{}'::jsonb)
-            )
-          END AS "latestEvent",
-          a.git_context AS "gitContext",
-          a.created_at AS "createdAt",
-          a.updated_at AS "updatedAt",
-          EXTRACT(EPOCH FROM (a.updated_at - a.created_at))::int * 1000 AS "durationMs",
-          COALESCE((
-            SELECT SUM(input_tokens + cache_creation_tokens + cache_read_tokens + output_tokens)
-            FROM agent_token_usage WHERE agent_id = a.id
-          ), 0)::bigint AS "totalTokens"
-         FROM agents a
-         ${whereClause}
-         ORDER BY ${sortSql}
-         LIMIT $${params.length + 1} OFFSET $${params.length + 2}`,
-        listParams
-      ),
-    ]);
-
-    const parentIds = agentsResult.rows.map((a: { id: string }) => a.id);
-
-    // Fetch child (persona/review) agents for these parents
-    type ChildAgent = {
-      id: string;
-      name: string;
-      persona: string | null;
-      status: string;
-      totalTokens: number;
-      createdAt: string;
-    };
-    const childrenByParent = new Map<string, ChildAgent[]>();
-
-    if (parentIds.length > 0) {
-      const childResult = await pool.query<
-        ChildAgent & { parentAgentId: string }
-      >(
-        `SELECT
-          a.id,
-          a.name,
-          a.persona,
-          a.status,
-          COALESCE((
-            SELECT SUM(input_tokens + cache_creation_tokens + cache_read_tokens + output_tokens)
-            FROM agent_token_usage WHERE agent_id = a.id
-          ), 0)::bigint AS "totalTokens",
-          a.created_at AS "createdAt",
-          a.parent_agent_id AS "parentAgentId"
-         FROM agents a
-         WHERE a.parent_agent_id = ANY($1)
-         ORDER BY a.created_at ASC`,
-        [parentIds]
-      );
-      for (const child of childResult.rows) {
-        const pid = child.parentAgentId;
-        let list = childrenByParent.get(pid);
-        if (!list) {
-          list = [];
-          childrenByParent.set(pid, list);
-        }
-        list.push({
-          id: child.id,
-          name: child.name,
-          persona: child.persona,
-          status: child.status,
-          totalTokens: child.totalTokens,
-          createdAt: child.createdAt,
-        });
-      }
-    }
-
-    const agents = agentsResult.rows.map(
-      (agent: { id: string; totalTokens: number }) => {
-        const children = childrenByParent.get(agent.id) ?? [];
-        const childTokens = children.reduce((sum, c) => sum + c.totalTokens, 0);
-        return {
-          ...agent,
-          children,
-          groupTotalTokens: agent.totalTokens + childTokens,
-        };
-      }
-    );
-
-    return { agents, total: countResult.rows[0]?.total ?? 0, limit, offset };
-  });
-
-  app.get("/api/v1/history/agents/:id", async (request, reply) => {
-    const { id } = request.params as { id: string };
-
-    const agentResult = await pool.query(
-      `SELECT
-        id, name, type, status, cwd,
-        worktree_path AS "worktreePath",
-        worktree_branch AS "worktreeBranch",
-        CASE
-          WHEN latest_event_type IS NULL OR latest_event_message IS NULL OR latest_event_updated_at IS NULL THEN NULL
-          ELSE json_build_object(
-            'type', latest_event_type,
-            'message', latest_event_message,
-            'updatedAt', latest_event_updated_at,
-            'metadata', COALESCE(latest_event_metadata, '{}'::jsonb)
-          )
-        END AS "latestEvent",
-        git_context AS "gitContext",
-        created_at AS "createdAt",
-        updated_at AS "updatedAt"
-       FROM agents WHERE id = $1`,
-      [id]
-    );
-    if (agentResult.rows.length === 0) {
-      return reply.code(404).send({ error: "Agent not found" });
-    }
-
-    const [
-      eventsResult,
-      tokenResult,
-      tokenByModelResult,
-      mediaResult,
-      feedbackResult,
-    ] = await Promise.all([
-      pool.query<{
-        id: number;
-        event_type: string;
-        message: string;
-        metadata: Record<string, unknown>;
-        created_at: string;
-      }>(
-        `SELECT id, event_type, message, metadata, created_at
-         FROM agent_events WHERE agent_id = $1 ORDER BY created_at ASC`,
-        [id]
-      ),
-      pool.query<{
-        total_input: number;
-        total_cache_creation: number;
-        total_cache_read: number;
-        total_output: number;
-        total_messages: number;
-      }>(
-        `SELECT
-          COALESCE(SUM(input_tokens), 0) AS total_input,
-          COALESCE(SUM(cache_creation_tokens), 0) AS total_cache_creation,
-          COALESCE(SUM(cache_read_tokens), 0) AS total_cache_read,
-          COALESCE(SUM(output_tokens), 0) AS total_output,
-          COALESCE(SUM(message_count), 0) AS total_messages
-         FROM agent_token_usage WHERE agent_id = $1`,
-        [id]
-      ),
-      pool.query<{
-        model: string;
-        input_tokens: number;
-        output_tokens: number;
-      }>(
-        `SELECT model,
-          SUM(input_tokens + cache_creation_tokens + cache_read_tokens) AS input_tokens,
-          SUM(output_tokens) AS output_tokens
-         FROM agent_token_usage WHERE agent_id = $1
-         GROUP BY model ORDER BY (SUM(input_tokens + cache_creation_tokens + cache_read_tokens) + SUM(output_tokens)) DESC`,
-        [id]
-      ),
-      pool.query<{
-        file_name: string;
-        source: string;
-        size_bytes: number;
-        description: string | null;
-        created_at: string;
-      }>(
-        `SELECT file_name, source, size_bytes, description, created_at
-         FROM media WHERE agent_id = $1 ORDER BY created_at`,
-        [id]
-      ),
-      pool.query<{
-        id: number;
-        agentId: string;
-        persona: string | null;
-        severity: string;
-        filePath: string | null;
-        lineNumber: number | null;
-        description: string;
-        suggestion: string | null;
-        mediaRef: string | null;
-        status: string;
-        createdAt: string;
-      }>(
-        `SELECT f.id, f.agent_id AS "agentId", a.persona, f.severity, f.file_path AS "filePath",
-                f.line_number AS "lineNumber", f.description, f.suggestion, f.media_ref AS "mediaRef",
-                f.status, f.created_at AS "createdAt"
-         FROM agent_feedback f
-         JOIN agents a ON a.id = f.agent_id
-         WHERE a.parent_agent_id = $1
-         ORDER BY f.created_at ASC
-         LIMIT 500`,
-        [id]
-      ),
-    ]);
-
-    const eventRows: ActivityEventRow[] = eventsResult.rows.map((r) => ({
-      agent_id: id,
-      event_type: r.event_type,
-      created_at: new Date(r.created_at),
-    }));
-    const stats = computeActivityStats(eventRows, null);
-
-    return {
-      agent: agentResult.rows[0],
-      events: eventsResult.rows,
-      tokenUsage: {
-        ...tokenResult.rows[0],
-        by_model: tokenByModelResult.rows,
-      },
-      media: mediaResult.rows,
-      feedback: feedbackResult.rows,
-      stateDurations: stats.stateDurations,
-    };
-  });
-
-  // ── Agents ────────────────────────────────────────────────────────
-
-  app.get("/api/v1/agents", async () => {
-    const agents = await agentManager.listAgents();
-    return { agents: agents.map(withStreamFlag) };
-  });
-
-  app.get("/api/v1/agents/git-context", async (request, reply) => {
-    const query = request.query as { ids?: unknown };
-    const ids =
-      typeof query.ids === "string"
-        ? query.ids
-            .split(",")
-            .map((id) => id.trim())
-            .filter((id) => id.length > 0)
-        : [];
-
-    const idFilter = ids.length > 0 ? new Set(ids) : null;
-    const agents = await agentManager.listAgents();
-    const targets = idFilter
-      ? agents.filter((agent) => idFilter.has(agent.id))
-      : agents;
-
-    const contexts = targets.map((agent) => ({
-      id: agent.id,
-      gitContext: agent.gitContext,
-    }));
-
-    return { contexts };
-  });
-
-  app.get("/api/v1/events", async (_request, reply) => {
-    reply.raw.setHeader("Content-Type", "text/event-stream");
-    reply.raw.setHeader("Cache-Control", "no-cache, no-transform");
-    reply.raw.setHeader("Connection", "keep-alive");
-    reply.raw.setHeader("X-Accel-Buffering", "no");
-    reply.hijack();
-
-    const stream = reply.raw;
-    const unsubscribe = uiEventBroker.subscribe(stream);
-    const heartbeat = setInterval(() => {
-      stream.write(": keepalive\n\n");
-    }, 20_000);
-
-    try {
-      const agents = await agentManager.listAgents();
-      uiEventBroker.sendSnapshot(stream, agents.map(withStreamFlag));
-    } catch (error) {
-      app.log.warn({ err: error }, "Failed to load SSE snapshot.");
-    }
-
-    stream.on("close", () => {
-      clearInterval(heartbeat);
-      unsubscribe();
-    });
-  });
-
-  app.post("/api/v1/notifications/ack", async (request, reply) => {
-    const body = request.body as { notificationId?: unknown };
-    if (typeof body?.notificationId !== "string") {
-      return reply
-        .code(400)
-        .send({ error: "notificationId must be a string." });
-    }
-    const found = ackWebNotification(body.notificationId);
-    app.log.debug(
-      { notificationId: body.notificationId, found },
-      "Web notification ack received"
-    );
-    return reply.code(204).send();
-  });
-
-  app.post("/api/v1/focus", async (request, reply) => {
-    const body = request.body as { agentId?: unknown };
-    const agentId = body?.agentId;
-
-    if (agentId === null || agentId === undefined) {
-      // User is no longer focused on any agent — clear all focus immediately
-      // so notifications resume without waiting for TTL expiry.
-      focusTracker.clearAll();
-      return reply.code(204).send();
-    }
-
-    if (typeof agentId !== "string" || !agentId.trim()) {
-      return reply
-        .code(400)
-        .send({ error: "agentId must be a non-empty string or null." });
-    }
-
-    focusTracker.setFocused(agentId.trim());
-    return reply.code(204).send();
-  });
-
-  app.get("/api/v1/agents/:id", async (request, reply) => {
-    const params = request.params as { id?: string };
-    const id = params.id ?? "";
-    const agent = await agentManager.getAgent(id);
-
-    if (!agent) {
-      return reply.code(404).send({ error: "Agent not found." });
-    }
-    return { agent: withStreamFlag(agent) };
-  });
-
-  app.patch("/api/v1/agents/:id/review-agent-type", async (request, reply) => {
-    const params = request.params as { id?: string };
-    const id = params.id ?? "";
-    const body = request.body as { reviewAgentType?: unknown } | null;
-
-    let reviewAgentType: (typeof CLI_AGENT_TYPES)[number] | null;
-    if (body?.reviewAgentType === null || body?.reviewAgentType === undefined) {
-      reviewAgentType = null;
-    } else if (
-      typeof body.reviewAgentType === "string" &&
-      CLI_AGENT_TYPES.includes(
-        body.reviewAgentType as (typeof CLI_AGENT_TYPES)[number]
-      )
-    ) {
-      reviewAgentType =
-        body.reviewAgentType as (typeof CLI_AGENT_TYPES)[number];
-    } else {
-      return reply.code(400).send({
-        error: `reviewAgentType must be null or one of ${CLI_AGENT_TYPES.join(", ")}.`,
-      });
-    }
-
-    if (reviewAgentType) {
-      const enabledAgentTypes = await getEnabledAgentTypes(pool);
-      if (!enabledAgentTypes.includes(reviewAgentType)) {
-        return reply.code(400).send({
-          error: `${reviewAgentType} agents are disabled in settings.`,
-        });
-      }
-    }
-
-    try {
-      await agentManager.updateReviewAgentType(id, reviewAgentType);
-      const agent = await agentManager.getAgent(id);
-      if (!agent) {
-        return reply.code(404).send({ error: "Agent not found." });
-      }
-      uiEventBroker.publish({
-        type: "agent.upsert",
-        agent: withStreamFlag(agent),
-      });
-      return { agent: withStreamFlag(agent) };
-    } catch (error) {
-      return handleAgentError(reply, error);
-    }
-  });
-
-  app.post("/api/v1/agents/:id/latest-event", async (request, reply) => {
-    const params = request.params as { id?: string };
-    const body = request.body as {
-      type?: unknown;
-      message?: unknown;
-      metadata?: unknown;
-    };
-    const id = params.id ?? "";
-    const type = body?.type;
-    const message = body?.message;
-    const metadata = body?.metadata;
-
-    if (!isAgentLatestEventType(type)) {
-      return reply.code(400).send({
-        error: `type must be one of: ${AGENT_LATEST_EVENT_TYPES.join(", ")}.`,
-      });
-    }
-
-    if (typeof message !== "string" || !message.trim()) {
-      return reply
-        .code(400)
-        .send({ error: "message must be a non-empty string." });
-    }
-
-    if (
-      metadata !== undefined &&
-      (metadata === null ||
-        typeof metadata !== "object" ||
-        Array.isArray(metadata))
-    ) {
-      return reply
-        .code(400)
-        .send({ error: "metadata must be an object when provided." });
-    }
-
-    const agent = await agentManager.upsertLatestEvent(id, {
-      type,
-      message: message.trim(),
-      metadata: metadata as Record<string, unknown> | undefined,
-    });
-
-    uiEventBroker.publish({
-      type: "agent.upsert",
-      agent: withStreamFlag(agent),
-    });
-    return { agent };
+  await registerJobRoutes(app, {
+    jobService,
+    publishUiEvent: (event) => uiEventBroker.publish(event as UiEvent),
+  });
+
+  await registerMcpRoutes(app, {
+    config,
+    agentManager,
+    jobService,
+    getBearerToken,
+    validateJobMcpToken,
+    validateAgentMcpToken,
+    resolveRepoRoot,
+    resolveWorktreeRoot,
+    mcpSendNotify,
+    mcpUpsertEvent,
+    mcpRenameSession,
+    mcpShareMedia,
+    mcpListMedia,
+    mcpSubmitFeedback,
+    mcpListPersonas,
+    mcpLaunchPersona,
+    mcpGetFeedback,
+    mcpResolveFeedback,
+    mcpSubmitResolution,
+    mcpCancelRecheck,
+    mcpUpsertPin,
+    mcpDeletePin,
+    mcpGetParentContext,
+    mcpGetRecheckContext,
+    mcpUpdateReviewStatus,
+    mcpCompleteReview,
+    mcpJobComplete,
+    mcpJobFailed,
+    mcpJobNeedsInput,
+    mcpJobLog,
+    mcpMethodNotAllowed,
+  });
+
+  await registerSystemRoutes(app, {
+    pool,
+    appLog: app.log,
+    slackNotifier,
+    iconColorKey: ICON_COLOR_KEY,
+    validIconColors: VALID_ICON_COLORS,
+    getCachedIconColor: () => cachedIconColor,
+    rewriteForColor: (color) => rewriteForColor(color as IconColor),
+    pendingGitRefreshEnqueuedAt,
+    gitRefreshDurationsMs,
+    gitRefreshAgentDiagnostics,
+    pendingGitRefreshAgentIds,
+    activeGitRefreshAgentIds,
+    gitRefreshCounters,
+    probeCommandTimeoutMs: PROBE_COMMAND_TIMEOUT_MS,
+    gitContextRefreshIntervalMs: GIT_CONTEXT_REFRESH_INTERVAL_MS,
+    gitContextRefreshConcurrency: GIT_CONTEXT_REFRESH_CONCURRENCY,
+    percentile,
+    toIso,
+  });
+
+  await registerActivityRoutes(app, {
+    pool,
+    agentManager,
+    parseActivityQuery,
+    loadScopedActivityEvents,
+    timeRangeClause,
+    dateTruncTz,
+    escapeLike,
+  });
+
+  await registerReleaseRoutes(app, {
+    pool,
+    appLog: app.log,
+    config,
+    serverDir,
+    agentManager,
+    worktreeLocationKey: WORKTREE_LOCATION_KEY,
+    validWorktreeLocations: VALID_WORKTREE_LOCATIONS,
+    getActiveReleaseJob: () => activeReleaseJob as RouteReleaseJob | null,
+    setActiveReleaseJob: (job) => {
+      activeReleaseJob = job as ReleaseJob | null;
+    },
+    getActiveAssistedUpdateLaunch: () => activeAssistedUpdateLaunch,
+    setActiveAssistedUpdateLaunch: (active) => {
+      activeAssistedUpdateLaunch = active;
+    },
+    releaseStreamClients,
+    getAppVersionInfo,
+    getGitHubRepo,
+    parseGhJson,
+    compareSemver,
+    checkIsAdmin,
+    fetchReleaseMetadata,
+    fetchLatestReleaseMetadata,
+    dispatchBaseUrl,
+    dispatchHealthUrl,
+    defaultServiceRestartCommand,
+    buildAssistedUpdatePrompt,
+    hasActiveAssistedUpdateAgent,
+    broadcastReleaseEvent: (event) =>
+      broadcastReleaseEvent(event as ReleaseStreamEvent),
+    appendReleaseLog: (job, line) => appendReleaseLog(job as ReleaseJob, line),
+    rehydrateActiveAssistedJob,
+    runReleaseJob: (job) => runReleaseJob(job as ReleaseJob),
+    runUpdateJob: (job) => runUpdateJob(job as ReleaseJob),
+    getBearerToken,
+    queueGitContextRefresh,
+    publishUiEvent: (event) => uiEventBroker.publish(event as UiEvent),
+    withStreamFlag,
+    handleAgentError,
   });
 
   await registerMediaRoutes(app, {
@@ -4183,527 +1808,60 @@ async function registerRoutes() {
     publishUiEvent: (event) => uiEventBroker.publish(event as UiEvent),
   });
 
-  app.post("/api/v1/agents/:id/stream", async (request, reply) => {
-    const params = request.params as { id?: string };
-    const id = params.id ?? "";
-    const agent = await agentManager.getAgent(id);
-    if (!agent) {
-      return reply.code(404).send({ error: "Agent not found." });
-    }
-
-    const body = request.body as {
-      type?: unknown;
-      port?: unknown;
-      description?: unknown;
-    };
-    if (body?.type === "stop") {
-      const description =
-        typeof body.description === "string" ? body.description : null;
-      streamManager.stopStream(id, description);
-      return { ok: true };
-    }
-
-    if (body?.type === "playwright") {
-      if (
-        typeof body.port !== "number" ||
-        !Number.isFinite(body.port) ||
-        body.port < 1
-      ) {
-        return reply
-          .code(400)
-          .send({ error: "port must be a positive number." });
-      }
-      if (streamManager.hasStream(id)) {
-        return reply
-          .code(409)
-          .send({ error: "Stream already active for this agent." });
-      }
-      try {
-        await streamManager.startStream(id, body.port);
-      } catch (error) {
-        const message =
-          error instanceof Error ? error.message : "Failed to start stream.";
-        return reply.code(502).send({ error: message });
-      }
-      return { ok: true };
-    }
-
-    return reply
-      .code(400)
-      .send({ error: "type must be 'playwright' or 'stop'." });
-  });
-
-  app.get("/api/v1/agents/:id/stream", async (request, reply) => {
-    const params = request.params as { id?: string };
-    const id = params.id ?? "";
-    const agent = await agentManager.getAgent(id);
-    if (!agent) {
-      return reply.code(404).send({ error: "Agent not found." });
-    }
-    if (!streamManager.hasStream(id)) {
-      return reply
-        .code(404)
-        .send({ error: "No active stream for this agent." });
-    }
-
-    reply.hijack();
-    const raw = reply.raw;
-    raw.writeHead(200, {
-      "Content-Type": "multipart/x-mixed-replace; boundary=frame",
-      "Cache-Control": "no-cache, no-transform",
-      Connection: "keep-alive",
-      "X-Accel-Buffering": "no",
-    });
-    raw.flushHeaders();
-    if (raw.socket) {
-      raw.socket.setNoDelay(true);
-    }
-
-    const unsubscribe = streamManager.addViewer(id, raw);
-    reply.raw.on("close", () => {
-      unsubscribe();
-    });
-  });
-
-  app.get("/api/v1/agents/:id/stream/viewer", async (request, reply) => {
-    const params = request.params as { id?: string };
-    const id = params.id ?? "";
-    const agent = await agentManager.getAgent(id);
-    if (!agent) {
-      return reply.code(404).send({ error: "Agent not found." });
-    }
-
-    const html = `<!DOCTYPE html>
-<html><head>
-<meta charset="utf-8">
-<title>${escapeHtml(agent.name)} — Live Stream</title>
-<style>
-  *{margin:0;padding:0;box-sizing:border-box}
-  body{background:#0a0a0a;display:flex;align-items:center;justify-content:center;height:100vh;overflow:hidden}
-  img{max-width:100%;max-height:100%;object-fit:contain}
-  .gone{display:flex;align-items:center;justify-content:center;height:100vh;color:#666;font-family:system-ui;font-size:14px}
-</style>
-</head><body>
-<img id="feed" src="/api/v1/agents/${escapeHtml(id)}/stream" alt="Live stream">
-<script>
-  const img = document.getElementById('feed');
-  img.onerror = () => {
-    document.body.innerHTML = '<div class="gone">Stream ended.</div>';
-  };
-</script>
-</body></html>`;
-    return reply.type("text/html").send(html);
-  });
-
-  app.post("/api/v1/agents", async (request, reply) => {
-    let parsedRequest: {
-      body: CreateAgentBody;
-      startupFiles: StartupFileUpload[];
-      isMultipart: boolean;
-    };
-    try {
-      parsedRequest = await parseCreateAgentRequest(request);
-    } catch (error) {
-      return reply.code(400).send({
-        error: error instanceof Error ? error.message : "Invalid request body.",
-      });
-    }
-    const { body, startupFiles } = parsedRequest;
-
-    if (typeof body?.cwd !== "string") {
-      return reply
-        .code(400)
-        .send({ error: "Body must include cwd as a string." });
-    }
-
-    let parsedAgentArgs: string[] | undefined;
-    let startupLinks: string[] | undefined;
-    let fullAccess: boolean | undefined;
-    let useWorktree: boolean | undefined;
-    let createNewBranch: boolean | undefined;
-    let autoReview: boolean | undefined;
-
-    try {
-      parsedAgentArgs = parseOptionalStringArrayField(
-        body.agentArgs ?? body.codexArgs,
-        "agentArgs",
-        parsedRequest.isMultipart
-      );
-      startupLinks = parseOptionalStringArrayField(
-        body.startupLinks,
-        "startupLinks",
-        parsedRequest.isMultipart
-      );
-      fullAccess = parseOptionalBooleanField(
-        body.fullAccess,
-        "fullAccess",
-        parsedRequest.isMultipart
-      );
-      useWorktree = parseOptionalBooleanField(
-        body.useWorktree,
-        "useWorktree",
-        parsedRequest.isMultipart
-      );
-      createNewBranch = parseOptionalBooleanField(
-        body.createNewBranch,
-        "createNewBranch",
-        parsedRequest.isMultipart
-      );
-      autoReview = parseOptionalBooleanField(
-        body.autoReview,
-        "autoReview",
-        parsedRequest.isMultipart
-      );
-    } catch (error) {
-      return reply.code(400).send({
-        error: error instanceof Error ? error.message : "Invalid request body.",
-      });
-    }
-
-    if (
-      body.type !== undefined &&
-      body.type !== "codex" &&
-      body.type !== "claude" &&
-      body.type !== "opencode" &&
-      body.type !== "terminal"
-    ) {
-      return reply.code(400).send({
-        error:
-          "type must be codex, claude, opencode, or terminal when provided.",
-      });
-    }
-
-    if (
-      body.worktreeBranch !== undefined &&
-      typeof body.worktreeBranch !== "string"
-    ) {
-      return reply
-        .code(400)
-        .send({ error: "worktreeBranch must be a string when provided." });
-    }
-
-    if (body.baseBranch !== undefined && typeof body.baseBranch !== "string") {
-      return reply
-        .code(400)
-        .send({ error: "baseBranch must be a string when provided." });
-    }
-
-    if (
-      body.initialPrompt !== undefined &&
-      typeof body.initialPrompt !== "string"
-    ) {
-      return reply
-        .code(400)
-        .send({ error: "initialPrompt must be a string when provided." });
-    }
-
-    if (
-      typeof body.initialPrompt === "string" &&
-      body.initialPrompt.trim().length > AGENT_INITIAL_PROMPT_MAX_CHARS
-    ) {
-      return reply.code(400).send({
-        error: `initialPrompt must be at most ${AGENT_INITIAL_PROMPT_MAX_CHARS} characters when provided.`,
-      });
-    }
-
-    const agentType =
-      body.type === "claude"
-        ? "claude"
-        : body.type === "opencode"
-          ? "opencode"
-          : body.type === "terminal"
-            ? "terminal"
-            : "codex";
-    const enabledAgentTypes = await getEnabledAgentTypes(pool);
-    if (!enabledAgentTypes.includes(agentType)) {
-      return reply
-        .code(400)
-        .send({ error: `${agentType} agents are disabled in settings.` });
-    }
-
-    // Terminal agents have no CLI to drive, so full-access / auto-review /
-    // initial-prompt are inert. Normalize them here so the stored record
-    // matches what the runtime honors and future side-effects keyed off these
-    // fields don't silently misfire for terminal agents.
-    const isTerminalAgent = agentType === "terminal";
-    const fullAccessArg =
-      agentType === "claude"
-        ? CLAUDE_FULL_ACCESS_ARG
-        : agentType === "codex"
-          ? CODEX_FULL_ACCESS_ARG
-          : null;
-    const resolvedAgentArgs =
-      !isTerminalAgent && fullAccess === true && fullAccessArg
-        ? Array.from(new Set([...(parsedAgentArgs ?? []), fullAccessArg]))
-        : parsedAgentArgs;
-    let startupPins: ReturnType<typeof createStartupPins>;
-    try {
-      startupPins = createStartupPins(startupLinks ?? []);
-    } catch (error) {
-      return reply.code(400).send({
-        error: error instanceof Error ? error.message : "Invalid startupLinks.",
-      });
-    }
-
-    try {
-      const worktreeLocationRaw = await getSetting(pool, WORKTREE_LOCATION_KEY);
-      const worktreeLocation: WorktreeLocation =
-        worktreeLocationRaw &&
-        (VALID_WORKTREE_LOCATIONS as string[]).includes(worktreeLocationRaw)
-          ? (worktreeLocationRaw as WorktreeLocation)
-          : "sibling";
-
-      const agent = await agentManager.createAgent({
-        name: typeof body.name === "string" ? body.name : undefined,
-        type: agentType,
-        cwd: body.cwd,
-        agentArgs: resolvedAgentArgs,
-        fullAccess: !isTerminalAgent && fullAccess === true,
-        useWorktree,
-        createNewBranch,
-        worktreeBranch:
-          typeof body.worktreeBranch === "string"
-            ? body.worktreeBranch
-            : undefined,
-        baseBranch:
-          typeof body.baseBranch === "string" ? body.baseBranch : undefined,
-        worktreeLocation,
-        persona: typeof body.persona === "string" ? body.persona : undefined,
-        parentAgentId:
-          typeof body.parentAgentId === "string"
-            ? body.parentAgentId
-            : undefined,
-        personaContext:
-          typeof body.personaContext === "string"
-            ? body.personaContext
-            : undefined,
-        autoReview: !isTerminalAgent && autoReview === true,
-        initialPrompt:
-          !isTerminalAgent && typeof body.initialPrompt === "string"
-            ? body.initialPrompt.trim() || undefined
-            : undefined,
-        initialPins: !isTerminalAgent ? startupPins : [],
-        initialFiles: !isTerminalAgent ? startupFiles : [],
-      });
-      queueGitContextRefresh([agent.id]);
-      uiEventBroker.publish({
-        type: "agent.upsert",
-        agent: withStreamFlag(agent),
-      });
-      return reply.code(201).send({ agent });
-    } catch (error) {
-      return handleAgentError(reply, error);
-    }
-  });
-
-  // Setup script callbacks — called by the bash setup script running in tmux
-  app.post("/api/v1/agents/:id/setup/error", async (request, reply) => {
-    const params = request.params as { id?: string };
-    const body = request.body as { message?: unknown };
-    const id = params.id ?? "";
-    const message =
-      typeof body?.message === "string" ? body.message : "Setup failed.";
-
-    try {
-      const agent = await agentManager.markSetupFailed(id, message);
-      uiEventBroker.publish({
-        type: "agent.upsert",
-        agent: withStreamFlag(agent),
-      });
-      return { ok: true };
-    } catch (error) {
-      return handleAgentError(reply, error);
-    }
-  });
-
-  app.post("/api/v1/agents/:id/setup/phase", async (request, reply) => {
-    const params = request.params as { id?: string };
-    const body = request.body as { phase?: unknown };
-    const id = params.id ?? "";
-
-    const validPhases = ["worktree", "env", "deps", "session"];
-    if (typeof body?.phase !== "string" || !validPhases.includes(body.phase)) {
-      return reply
-        .code(400)
-        .send({ error: "phase must be one of: worktree, env, deps, session" });
-    }
-
-    try {
-      await agentManager.updateSetupPhase(
-        id,
-        body.phase as "worktree" | "env" | "deps" | "session"
-      );
-      const agent = await agentManager.getAgent(id);
-      if (agent) {
+  await registerAgentRoutes(app, {
+    pool,
+    appLog: app.log,
+    agentManager,
+    worktreeLocationKey: WORKTREE_LOCATION_KEY,
+    validWorktreeLocations: VALID_WORKTREE_LOCATIONS,
+    queueGitContextRefresh,
+    publishUiEvent: (event) => uiEventBroker.publish(event as UiEvent),
+    subscribeUiEvents: (stream) => uiEventBroker.subscribe(stream),
+    sendUiSnapshot: (stream, agents) =>
+      uiEventBroker.sendSnapshot(stream, agents),
+    ackWebNotification,
+    clearFocusedAgents: () => focusTracker.clearAll(),
+    setFocusedAgent: (agentId) => focusTracker.setFocused(agentId),
+    withStreamFlag,
+    handleAgentError,
+    startStream: (agentId, port) => streamManager.startStream(agentId, port),
+    stopStream: (agentId, description) =>
+      streamManager.stopStream(agentId, description),
+    hasStream: (agentId) => streamManager.hasStream(agentId),
+    addStreamViewer: (agentId, stream) =>
+      streamManager.addViewer(agentId, stream),
+    issueTerminalToken: (agentId) => terminalTokenStore.issue(agentId),
+    consumeTerminalToken: (agentId, token) =>
+      terminalTokenStore.consume(agentId, token),
+    onArchivedAgentsDeleted: (deletedIds) => {
+      for (const deletedId of deletedIds) {
+        streamManager.stopStream(deletedId);
+        pendingGitRefreshAgentIds.delete(deletedId);
+        pendingGitRefreshEnqueuedAt.delete(deletedId);
+        activeGitRefreshAgentIds.delete(deletedId);
+        gitRefreshAgentDiagnostics.delete(deletedId);
         uiEventBroker.publish({
-          type: "agent.upsert",
-          agent: withStreamFlag(agent),
+          type: "agent.deleted",
+          agentId: deletedId,
         });
       }
-      return { ok: true };
-    } catch (error) {
-      return handleAgentError(reply, error);
-    }
-  });
-
-  app.post("/api/v1/agents/:id/setup/complete", async (request, reply) => {
-    const params = request.params as { id?: string };
-    const body = request.body as {
-      effectiveCwd?: unknown;
-      worktreePath?: unknown;
-      worktreeBranch?: unknown;
-    };
-    const id = params.id ?? "";
-
-    if (typeof body?.effectiveCwd !== "string") {
-      return reply.code(400).send({ error: "effectiveCwd must be a string." });
-    }
-
-    try {
-      const agent = await agentManager.completeSetup(id, {
-        effectiveCwd: body.effectiveCwd,
-        worktreePath:
-          typeof body.worktreePath === "string" ? body.worktreePath : null,
-        worktreeBranch:
-          typeof body.worktreeBranch === "string" ? body.worktreeBranch : null,
-      });
-      queueGitContextRefresh([agent.id]);
-      uiEventBroker.publish({
-        type: "agent.upsert",
-        agent: withStreamFlag(agent),
-      });
-      return { ok: true };
-    } catch (error) {
-      return handleAgentError(reply, error);
-    }
-  });
-
-  app.post("/api/v1/agents/:id/start", async (request, reply) => {
-    const params = request.params as { id?: string };
-    const id = params.id ?? "";
-
-    try {
-      const agent = await agentManager.startAgent(id);
-      queueGitContextRefresh([agent.id]);
-      uiEventBroker.publish({
-        type: "agent.upsert",
-        agent: withStreamFlag(agent),
-      });
-      return { agent };
-    } catch (error) {
-      return handleAgentError(reply, error);
-    }
-  });
-
-  app.post("/api/v1/agents/:id/stop", async (request, reply) => {
-    const params = request.params as { id?: string };
-    const body = request.body as { force?: unknown } | undefined;
-    const id = params.id ?? "";
-
-    app.log.info(
-      { agentId: id, force: body?.force ?? false },
-      "Stop agent requested"
-    );
-
-    if (body?.force !== undefined && typeof body.force !== "boolean") {
-      return reply
-        .code(400)
-        .send({ error: "force must be a boolean when provided." });
-    }
-
-    try {
-      const agent = await agentManager.stopAgent(id, {
-        force: body?.force as boolean | undefined,
-      });
-      queueGitContextRefresh([agent.id]);
-      uiEventBroker.publish({
-        type: "agent.upsert",
-        agent: withStreamFlag(agent),
-      });
-      return { agent };
-    } catch (error) {
-      return handleAgentError(reply, error);
-    }
-  });
-
-  app.get("/api/v1/agents/:id/worktree-status", async (request, reply) => {
-    const params = request.params as { id?: string };
-    const id = params.id ?? "";
-
-    try {
-      const status = await agentManager.checkWorktreeStatus(id);
-      return status;
-    } catch (error) {
-      return handleAgentError(reply, error);
-    }
-  });
-
-  app.delete("/api/v1/agents/:id", async (request, reply) => {
-    const params = request.params as { id?: unknown };
-    const query = request.query as { cleanupWorktree?: unknown };
-
-    if (typeof params.id !== "string") {
-      return reply.code(400).send({ error: "Missing agent id." });
-    }
-
-    const validCleanupModes = ["auto", "keep", "force"] as const;
-    type CleanupMode = (typeof validCleanupModes)[number];
-    const cleanupWorktree: CleanupMode =
-      typeof query.cleanupWorktree === "string" &&
-      (validCleanupModes as readonly string[]).includes(query.cleanupWorktree)
-        ? (query.cleanupWorktree as CleanupMode)
-        : "auto";
-
-    try {
-      // Fast synchronous phase: mark agent as archiving and return immediately.
-      // cleanupWorktree is persisted so reconciliation can honor it if the server restarts.
-      const agent = await agentManager.beginArchive(params.id, cleanupWorktree);
-      uiEventBroker.publish({
-        type: "agent.upsert",
-        agent: withStreamFlag(agent),
-      });
-
-      // Fire-and-forget: run cleanup in background (tracked for graceful shutdown + dedup)
-      const agentId = params.id;
+    },
+    onArchiveError: (agentId, error) => {
+      app.log.error({ err: error, agentId }, "Background archive failed");
+    },
+    trackArchivePromise: (agentId, archivePromise) => {
       archivingAgentIds.add(agentId);
-      const archivePromise = agentManager.executeArchive(agentId, {
-        onPhaseChange: (updated) => {
-          uiEventBroker.publish({
-            type: "agent.upsert",
-            agent: withStreamFlag(updated),
-          });
-        },
-        onComplete: (deletedIds) => {
-          for (const deletedId of deletedIds) {
-            streamManager.stopStream(deletedId);
-            pendingGitRefreshAgentIds.delete(deletedId);
-            pendingGitRefreshEnqueuedAt.delete(deletedId);
-            activeGitRefreshAgentIds.delete(deletedId);
-            gitRefreshAgentDiagnostics.delete(deletedId);
-            uiEventBroker.publish({
-              type: "agent.deleted",
-              agentId: deletedId,
-            });
-          }
-        },
-        onError: (error) => {
-          app.log.error({ err: error, agentId }, "Background archive failed");
-        },
-      });
       activeArchives.add(archivePromise);
       archivePromise.finally(() => {
         activeArchives.delete(archivePromise);
         archivingAgentIds.delete(agentId);
       });
-
-      return reply.code(202).send({ status: "archiving" });
-    } catch (error) {
-      return handleAgentError(reply, error);
-    }
+    },
   });
 
   // --- Personas ---
   await registerPersonaReviewRoutes(app, {
+    pool,
     agentManager,
     resolveWorktreeRoot,
     resolveRepoRoot,
@@ -4723,146 +1881,6 @@ async function registerRoutes() {
     publishUiEvent: (event) => uiEventBroker.publish(event as UiEvent),
     handleAgentError,
   });
-
-  app.post("/api/v1/agents/:id/terminal/token", async (request, reply) => {
-    const params = request.params as { id?: string };
-    const id = params.id ?? "";
-
-    try {
-      const access = await agentManager.getTerminalAccess(id);
-      if (access.mode === "inert") {
-        return {
-          mode: "inert" as const,
-          message: access.message,
-        };
-      }
-      const token = terminalTokenStore.issue(id);
-      return {
-        mode: "tmux" as const,
-        token,
-        wsUrl: `/api/v1/agents/${id}/terminal/ws?token=${token}`,
-      };
-    } catch (error) {
-      // Keep UI state in sync when terminal access lookup corrected a stale running status.
-      const refreshed = await agentManager.getAgent(id);
-      if (refreshed) {
-        queueGitContextRefresh([refreshed.id]);
-        uiEventBroker.publish({
-          type: "agent.upsert",
-          agent: withStreamFlag(refreshed),
-        });
-      }
-      return handleAgentError(reply, error);
-    }
-  });
-
-  app.get(
-    "/api/v1/agents/:id/terminal/ws",
-    { websocket: true },
-    async (socket, request) => {
-      const params = request.params as { id?: string };
-      const query = request.query as {
-        token?: string;
-        cols?: string;
-        rows?: string;
-      };
-      const agentId = params.id ?? "";
-      const token = query.token ?? "";
-
-      if (!terminalTokenStore.consume(agentId, token)) {
-        socket.send(
-          JSON.stringify({
-            type: "error",
-            message: "Invalid or expired terminal token.",
-          })
-        );
-        socket.close(1008, "invalid token");
-        return;
-      }
-
-      let tmuxSession: string;
-      try {
-        const access = await agentManager.getTerminalAccess(agentId);
-        if (access.mode !== "tmux") {
-          throw new Error(access.message);
-        }
-        tmuxSession = access.sessionName;
-        await runCommand(
-          "tmux",
-          ["set-option", "-t", tmuxSession, "mouse", "on"],
-          {
-            allowedExitCodes: [0],
-          }
-        );
-      } catch (error) {
-        const message =
-          error instanceof Error ? error.message : "Terminal attach failed.";
-        socket.send(JSON.stringify({ type: "error", message }));
-        socket.close(1011, "attach failed");
-        return;
-      }
-
-      const cols = Number(query.cols ?? 140);
-      const rows = Number(query.rows ?? 42);
-      const ptyProcess = spawnPty(
-        "tmux",
-        ["attach-session", "-t", tmuxSession],
-        {
-          name: "xterm-256color",
-          cols: Number.isFinite(cols) ? cols : 140,
-          rows: Number.isFinite(rows) ? rows : 42,
-        }
-      );
-
-      const sendJson = (payload: unknown): void => {
-        if (socket.readyState === socket.OPEN) {
-          socket.send(JSON.stringify(payload));
-        }
-      };
-
-      ptyProcess.onData((data) => {
-        sendJson({ type: "output", data });
-      });
-
-      ptyProcess.onExit((event) => {
-        sendJson({ type: "exit", exitCode: event.exitCode });
-        socket.close(1000, "terminal exited");
-      });
-
-      const heartbeatTimer = setInterval(() => {
-        sendJson({ type: "heartbeat", ts: Date.now() });
-      }, 20_000);
-
-      socket.on("message", (buffer) => {
-        const message = decodeClientMessage(buffer);
-        if (!message) {
-          sendJson({ type: "error", message: "Invalid message payload." });
-          return;
-        }
-
-        if (message.type === "input") {
-          if (!message.data) {
-            return;
-          }
-          ptyProcess.write(message.data);
-          return;
-        }
-
-        if (message.type === "resize") {
-          if (message.cols > 0 && message.rows > 0) {
-            ptyProcess.resize(message.cols, message.rows);
-          }
-        }
-      });
-
-      socket.on("close", () => {
-        clearInterval(heartbeatTimer);
-        try {
-          ptyProcess.kill();
-        } catch {}
-      });
-    }
-  );
 }
 
 async function waitForDatabase(maxAttempts = 15, delayMs = 2000) {
